factor out 700 line start command and callees
