+++ plugins/hypervisors/vmware/src/main/java/com/cloud/hypervisor/vmware/resource/VmwareResource.java
@@ -44,13 +44,10 @@ import java.util.UUID;
 import javax.naming.ConfigurationException;
 import javax.xml.datatype.XMLGregorianCalendar;
 
-import org.apache.cloudstack.api.ApiConstants;
 import org.apache.cloudstack.storage.command.CopyCommand;
 import org.apache.cloudstack.storage.command.StorageSubSystemCommand;
-import org.apache.cloudstack.storage.configdrive.ConfigDrive;
 import org.apache.cloudstack.storage.resource.NfsSecondaryStorageResource;
 import org.apache.cloudstack.storage.to.PrimaryDataStoreTO;
-import org.apache.cloudstack.storage.to.TemplateObjectTO;
 import org.apache.cloudstack.storage.to.VolumeObjectTO;
 import org.apache.cloudstack.utils.volume.VirtualMachineDiskInfo;
 import org.apache.cloudstack.vm.UnmanagedInstanceTO;
@@ -181,7 +178,6 @@ import com.cloud.agent.api.storage.PrimaryStorageDownloadCommand;
 import com.cloud.agent.api.storage.ResizeVolumeAnswer;
 import com.cloud.agent.api.storage.ResizeVolumeCommand;
 import com.cloud.agent.api.to.DataStoreTO;
-import com.cloud.agent.api.to.DiskTO;
 import com.cloud.agent.api.to.IpAddressTO;
 import com.cloud.agent.api.to.NfsTO;
 import com.cloud.agent.api.to.NicTO;
@@ -215,7 +211,6 @@ import com.cloud.hypervisor.vmware.mo.HostMO;
 import com.cloud.hypervisor.vmware.mo.HostStorageSystemMO;
 import com.cloud.hypervisor.vmware.mo.HypervisorHostHelper;
 import com.cloud.hypervisor.vmware.mo.NetworkDetails;
-import com.cloud.hypervisor.vmware.mo.TaskMO;
 import com.cloud.hypervisor.vmware.mo.VirtualEthernetCardType;
 import com.cloud.hypervisor.vmware.mo.VirtualMachineDiskInfoBuilder;
 import com.cloud.hypervisor.vmware.mo.VirtualMachineMO;
@@ -254,7 +249,6 @@ import com.cloud.utils.exception.ExceptionUtil;
 import com.cloud.utils.mgmt.JmxUtil;
 import com.cloud.utils.mgmt.PropertyMapDynamicBean;
 import com.cloud.utils.net.NetUtils;
-import com.cloud.utils.nicira.nvp.plugin.NiciraNvpApiVersion;
 import com.cloud.utils.script.Script;
 import com.cloud.utils.ssh.SshHelper;
 import com.cloud.vm.VirtualMachine;
@@ -265,12 +259,8 @@ import com.google.common.base.Strings;
 import com.google.gson.Gson;
 import com.vmware.vim25.AboutInfo;
 import com.vmware.vim25.ArrayUpdateOperation;
-import com.vmware.vim25.BoolPolicy;
 import com.vmware.vim25.ComputeResourceSummary;
 import com.vmware.vim25.CustomFieldStringValue;
-import com.vmware.vim25.DVPortConfigInfo;
-import com.vmware.vim25.DVPortConfigSpec;
-import com.vmware.vim25.DasVmPriority;
 import com.vmware.vim25.DatastoreInfo;
 import com.vmware.vim25.DatastoreSummary;
 import com.vmware.vim25.DistributedVirtualPort;
@@ -314,11 +304,9 @@ import com.vmware.vim25.VirtualDiskFlatVer2BackingInfo;
 import com.vmware.vim25.VirtualEthernetCard;
 import com.vmware.vim25.VirtualEthernetCardDistributedVirtualPortBackingInfo;
 import com.vmware.vim25.VirtualEthernetCardNetworkBackingInfo;
-import com.vmware.vim25.VirtualEthernetCardOpaqueNetworkBackingInfo;
 import com.vmware.vim25.VirtualIDEController;
 import com.vmware.vim25.VirtualMachineConfigSpec;
 import com.vmware.vim25.VirtualMachineBootOptions;
-import com.vmware.vim25.VirtualMachineFileInfo;
 import com.vmware.vim25.VirtualMachineFileLayoutEx;
 import com.vmware.vim25.VirtualMachineFileLayoutExFileInfo;
 import com.vmware.vim25.VirtualMachineGuestOsIdentifier;
@@ -330,7 +318,6 @@ import com.vmware.vim25.VirtualMachineToolsStatus;
 import com.vmware.vim25.VirtualMachineVideoCard;
 import com.vmware.vim25.VirtualPCNet32;
 import com.vmware.vim25.VirtualSCSIController;
-import com.vmware.vim25.VirtualUSBController;
 import com.vmware.vim25.VirtualVmxnet2;
 import com.vmware.vim25.VirtualVmxnet3;
 import com.vmware.vim25.VmConfigInfo;
@@ -344,12 +331,14 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
     public static final String VMDK_EXTENSION = ".vmdk";
 
     private static final Random RANDOM = new Random(System.nanoTime());
+    private final StartCommandExecutor startCommandExecutor = new StartCommandExecutor(this);
+
+    // FR37 Does this need to be a setting?
+    protected final int shutdownWaitMs = 300000;  // wait up to 5 minutes for shutdown
 
     protected String _name;
 
     protected final long _opsTimeout = 900000;   // 15 minutes time out to time
-
-    protected final int _shutdownWaitMs = 300000;  // wait up to 5 minutes for shutdown
 
     // out an operation
     protected final int _retry = 24;
@@ -358,7 +347,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
     protected final int MazCmdMBean = 100;
 
     protected String _url;
-    protected String _dcId;
+    protected String dcId;
     protected String _pod;
     protected String _cluster;
     protected String _username;
@@ -373,7 +362,8 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
     protected Map<String, String> _vsmCredentials = null;
     protected int _portsPerDvPortGroup;
     protected boolean _fullCloneFlag = false;
-    protected boolean _instanceNameFlag = false;
+    // FR37 move to global seting(s)
+    protected static boolean instanceNameFlag = false;
 
     protected boolean _recycleHungWorker = false;
     protected DiskControllerType _rootDiskController = DiskControllerType.ide;
@@ -384,7 +374,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected List<PropertyMapDynamicBean> _cmdMBeans = new ArrayList<PropertyMapDynamicBean>();
 
-    protected Gson _gson;
+    protected Gson gson;
 
     protected volatile long _cmdSequence = 1;
 
@@ -406,12 +396,24 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
     protected static final String s_relativePathSystemVmKeyFileInstallDir = "scripts/vm/systemvm/id_rsa.cloud";
     protected static final String s_defaultPathSystemVmKeyFile = "/usr/share/cloudstack-common/scripts/vm/systemvm/id_rsa.cloud";
 
-    public Gson getGson() {
-        return _gson;
+    public VmwareResource() {
+        gson = GsonHelper.getGsonLogger();
     }
 
-    public VmwareResource() {
-        _gson = GsonHelper.getGsonLogger();
+    public VmwareStorageProcessor getStorageProcessor() {
+        return _storageProcessor;
+    }
+
+    public int getShutdownWaitMs() {
+        return shutdownWaitMs;
+    }
+
+    public String getDcId() {
+        return dcId;
+    }
+
+    public Gson getGson() {
+        return gson;
     }
 
     private String getCommandLogTitle(Command cmd) {
@@ -441,7 +443,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
             Date startTime = DateUtil.currentGMTTime();
             PropertyMapDynamicBean mbean = new PropertyMapDynamicBean();
             mbean.addProp("StartTime", DateUtil.getDateDisplayString(TimeZone.getDefault(), startTime));
-            mbean.addProp("Command", _gson.toJson(cmd));
+            mbean.addProp("Command", gson.toJson(cmd));
             mbean.addProp("Sequence", String.valueOf(cmdSequence));
             mbean.addProp("Name", cmd.getClass().getSimpleName());
 
@@ -527,7 +529,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
             } else if (clz == NetworkUsageCommand.class) {
                 answer = execute((NetworkUsageCommand) cmd);
             } else if (clz == StartCommand.class) {
-                answer = execute((StartCommand) cmd);
+                answer = startCommandExecutor.execute((StartCommand)cmd);
             } else if (clz == CheckSshCommand.class) {
                 answer = execute((CheckSshCommand) cmd);
             } else if (clz == CheckNetworkCommand.class) {
@@ -571,7 +573,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
             Date doneTime = DateUtil.currentGMTTime();
             mbean.addProp("DoneTime", DateUtil.getDateDisplayString(TimeZone.getDefault(), doneTime));
-            mbean.addProp("Answer", _gson.toJson(answer));
+            mbean.addProp("Answer", gson.toJson(answer));
 
             synchronized (this) {
                 try {
@@ -632,7 +634,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         if (success) {
             s_logger.info("VmwareStorageProcessor and VmwareStorageSubsystemCommandHandler successfully reconfigured");
         } else {
-            s_logger.error("Error while reconfiguring VmwareStorageProcessor and VmwareStorageSubsystemCommandHandler, params=" + _gson.toJson(params));
+            s_logger.error("Error while reconfiguring VmwareStorageProcessor and VmwareStorageSubsystemCommandHandler, params=" + gson.toJson(params));
         }
     }
 
@@ -690,33 +692,6 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
             return true;
         }
         return false;
-    }
-
-    /**
-     * Registers the vm to the inventory given the vmx file.
-     */
-    private void registerVm(String vmName, DatastoreMO dsMo) throws Exception {
-
-        //1st param
-        VmwareHypervisorHost hyperHost = getHyperHost(getServiceContext());
-        ManagedObjectReference dcMor = hyperHost.getHyperHostDatacenter();
-        DatacenterMO dataCenterMo = new DatacenterMO(getServiceContext(), dcMor);
-        ManagedObjectReference vmFolderMor = dataCenterMo.getVmFolder();
-
-        //2nd param
-        String vmxFilePath = dsMo.searchFileInSubFolders(vmName + ".vmx", false, VmwareManager.s_vmwareSearchExcludeFolder.value());
-
-        // 5th param
-        ManagedObjectReference morPool = hyperHost.getHyperHostOwnerResourcePool();
-
-        ManagedObjectReference morTask = getServiceContext().getService().registerVMTask(vmFolderMor, vmxFilePath, vmName, false, morPool, hyperHost.getMor());
-        boolean result = getServiceContext().getVimClient().waitForTask(morTask);
-        if (!result) {
-            throw new Exception("Unable to register vm due to " + TaskMO.getTaskFailureInfo(getServiceContext(), morTask));
-        } else {
-            getServiceContext().waitForTaskProgressDone(morTask);
-        }
-
     }
 
     private Answer execute(ResizeVolumeCommand cmd) {
@@ -877,7 +852,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(CheckNetworkCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource CheckNetworkCommand " + _gson.toJson(cmd));
+            s_logger.info("Executing resource CheckNetworkCommand " + gson.toJson(cmd));
         }
 
         // TODO setup portgroup for private network needs to be done here now
@@ -889,7 +864,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
             return VPCNetworkUsage(cmd);
         }
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource NetworkUsageCommand " + _gson.toJson(cmd));
+            s_logger.info("Executing resource NetworkUsageCommand " + gson.toJson(cmd));
         }
         if (cmd.getOption() != null && cmd.getOption().equals("create")) {
             String result = networkUsage(cmd.getPrivateIP(), "create", null);
@@ -1134,7 +1109,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     private PlugNicAnswer execute(PlugNicCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource PlugNicCommand " + _gson.toJson(cmd));
+            s_logger.info("Executing resource PlugNicCommand " + gson.toJson(cmd));
         }
 
         getServiceContext().getStockObject(VmwareManager.CONTEXT_STOCK_NAME);
@@ -1218,7 +1193,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     private ReplugNicAnswer execute(ReplugNicCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource ReplugNicCommand " + _gson.toJson(cmd));
+            s_logger.info("Executing resource ReplugNicCommand " + gson.toJson(cmd));
         }
 
         getServiceContext().getStockObject(VmwareManager.CONTEXT_STOCK_NAME);
@@ -1298,7 +1273,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     private UnPlugNicAnswer execute(UnPlugNicCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource UnPlugNicCommand " + _gson.toJson(cmd));
+            s_logger.info("Executing resource UnPlugNicCommand " + gson.toJson(cmd));
         }
 
         VmwareContext context = getServiceContext();
@@ -1566,40 +1541,6 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         return new CheckSshAnswer(cmd);
     }
 
-    private DiskTO[] validateDisks(DiskTO[] disks) {
-        List<DiskTO> validatedDisks = new ArrayList<DiskTO>();
-
-        for (DiskTO vol : disks) {
-            if (vol.getType() != Volume.Type.ISO) {
-                VolumeObjectTO volumeTO = (VolumeObjectTO) vol.getData();
-                DataStoreTO primaryStore = volumeTO.getDataStore();
-                if (primaryStore.getUuid() != null && !primaryStore.getUuid().isEmpty()) {
-                    validatedDisks.add(vol);
-                }
-            } else if (vol.getType() == Volume.Type.ISO) {
-                TemplateObjectTO templateTO = (TemplateObjectTO) vol.getData();
-                if (templateTO.getPath() != null && !templateTO.getPath().isEmpty()) {
-                    validatedDisks.add(vol);
-                }
-            } else {
-                if (s_logger.isDebugEnabled()) {
-                    s_logger.debug("Drop invalid disk option, volumeTO: " + _gson.toJson(vol));
-                }
-            }
-        }
-        Collections.sort(validatedDisks, (d1, d2) -> d1.getDiskSeq().compareTo(d2.getDiskSeq()));
-        return validatedDisks.toArray(new DiskTO[0]);
-    }
-
-    private static DiskTO getIsoDiskTO(DiskTO[] disks) {
-        for (DiskTO vol : disks) {
-            if (vol.getType() == Volume.Type.ISO) {
-                return vol;
-            }
-        }
-        return null;
-    }
-
     protected ScaleVmAnswer execute(ScaleVmCommand cmd) {
 
         VmwareContext context = getServiceContext();
@@ -1647,798 +1588,17 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         return new ScaleVmAnswer(cmd, true, null);
     }
 
-    protected void ensureDiskControllers(VirtualMachineMO vmMo, Pair<String, String> controllerInfo) throws Exception {
-        if (vmMo == null) {
-            return;
-        }
-
-        String msg;
-        String rootDiskController = controllerInfo.first();
-        String dataDiskController = controllerInfo.second();
-        String scsiDiskController;
-        String recommendedDiskController = null;
-
-        if (VmwareHelper.isControllerOsRecommended(dataDiskController) || VmwareHelper.isControllerOsRecommended(rootDiskController)) {
-            recommendedDiskController = vmMo.getRecommendedDiskController(null);
-        }
-        scsiDiskController = HypervisorHostHelper.getScsiController(new Pair<String, String>(rootDiskController, dataDiskController), recommendedDiskController);
-        if (scsiDiskController == null) {
-            return;
-        }
-
-        vmMo.getScsiDeviceControllerKeyNoException();
-        // This VM needs SCSI controllers.
-        // Get count of existing scsi controllers. Helps not to attempt to create more than the maximum allowed 4
-        // Get maximum among the bus numbers in use by scsi controllers. Safe to pick maximum, because we always go sequential allocating bus numbers.
-        Ternary<Integer, Integer, DiskControllerType> scsiControllerInfo = vmMo.getScsiControllerInfo();
-        int requiredNumScsiControllers = VmwareHelper.MAX_SCSI_CONTROLLER_COUNT - scsiControllerInfo.first();
-        int availableBusNum = scsiControllerInfo.second() + 1; // method returned current max. bus number
-
-        if (requiredNumScsiControllers == 0) {
-            return;
-        }
-        if (scsiControllerInfo.first() > 0) {
-            // For VMs which already have a SCSI controller, do NOT attempt to add any more SCSI controllers & return the sub type.
-            // For Legacy VMs would have only 1 LsiLogic Parallel SCSI controller, and doesn't require more.
-            // For VMs created post device ordering support, 4 SCSI subtype controllers are ensured during deployment itself. No need to add more.
-            // For fresh VM deployment only, all required controllers should be ensured.
-            return;
-        }
-        ensureScsiDiskControllers(vmMo, scsiDiskController, requiredNumScsiControllers, availableBusNum);
-    }
-
-    private void ensureScsiDiskControllers(VirtualMachineMO vmMo, String scsiDiskController, int requiredNumScsiControllers, int availableBusNum) throws Exception {
-        // Pick the sub type of scsi
-        if (DiskControllerType.getType(scsiDiskController) == DiskControllerType.pvscsi) {
-            if (!vmMo.isPvScsiSupported()) {
-                String msg = "This VM doesn't support Vmware Paravirtual SCSI controller for virtual disks, because the virtual hardware version is less than 7.";
-                throw new Exception(msg);
-            }
-            vmMo.ensurePvScsiDeviceController(requiredNumScsiControllers, availableBusNum);
-        } else if (DiskControllerType.getType(scsiDiskController) == DiskControllerType.lsisas1068) {
-            vmMo.ensureLsiLogicSasDeviceControllers(requiredNumScsiControllers, availableBusNum);
-        } else if (DiskControllerType.getType(scsiDiskController) == DiskControllerType.buslogic) {
-            vmMo.ensureBusLogicDeviceControllers(requiredNumScsiControllers, availableBusNum);
-        } else if (DiskControllerType.getType(scsiDiskController) == DiskControllerType.lsilogic) {
-            vmMo.ensureLsiLogicDeviceControllers(requiredNumScsiControllers, availableBusNum);
-        }
-    }
-
     protected StartAnswer execute(StartCommand cmd) {
-        if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource StartCommand: " + _gson.toJson(cmd));
-        }
 
         // FR37 if startcommand contains a secendary storage URL or some flag or other type of indicator, deploy OVA as is
-        String secStorUrl = cmd.getSecondaryStorage();
-        if (StringUtils.isNotEmpty(secStorUrl)) {
-            if (s_logger.isTraceEnabled()) {
-                s_logger.trace(String.format("deploying OVA as is from %s", secStorUrl));
-            }
-        }
-        VirtualMachineTO vmSpec = cmd.getVirtualMachine();
-        boolean vmAlreadyExistsInVcenter = false;
-
-        String existingVmName = null;
-        VirtualMachineFileInfo existingVmFileInfo = null;
-        VirtualMachineFileLayoutEx existingVmFileLayout = null;
-        List<DatastoreMO> existingDatastores = new ArrayList<DatastoreMO>();
-
-        Pair<String, String> names = composeVmNames(vmSpec);
-        String vmInternalCSName = names.first();
-        String vmNameOnVcenter = names.second();
-        String dataDiskController = vmSpec.getDetails().get(VmDetailConstants.DATA_DISK_CONTROLLER);
-        String rootDiskController = vmSpec.getDetails().get(VmDetailConstants.ROOT_DISK_CONTROLLER);
-        DiskTO rootDiskTO = null;
-        String bootMode = null;
-        if (vmSpec.getDetails().containsKey(VmDetailConstants.BOOT_MODE)) {
-            bootMode = vmSpec.getDetails().get(VmDetailConstants.BOOT_MODE);
-        }
-        if (null == bootMode) {
-            bootMode = ApiConstants.BootType.BIOS.toString();
-        }
 
         // If root disk controller is scsi, then data disk controller would also be scsi instead of using 'osdefault'
         // This helps avoid mix of different scsi subtype controllers in instance.
-        if (DiskControllerType.osdefault == DiskControllerType.getType(dataDiskController) && DiskControllerType.lsilogic == DiskControllerType.getType(rootDiskController)) {
-            dataDiskController = DiskControllerType.scsi.toString();
-        }
 
         // Validate the controller types
-        dataDiskController = DiskControllerType.getType(dataDiskController).toString();
-        rootDiskController = DiskControllerType.getType(rootDiskController).toString();
 
-        if (DiskControllerType.getType(rootDiskController) == DiskControllerType.none) {
-            throw new CloudRuntimeException("Invalid root disk controller detected : " + rootDiskController);
-        }
-        if (DiskControllerType.getType(dataDiskController) == DiskControllerType.none) {
-            throw new CloudRuntimeException("Invalid data disk controller detected : " + dataDiskController);
-        }
-
-        Pair<String, String> controllerInfo = new Pair<String, String>(rootDiskController, dataDiskController);
-
-        Boolean systemVm = vmSpec.getType().isUsedBySystem();
         // Thus, vmInternalCSName always holds i-x-y, the cloudstack generated internal VM name.
-        VmwareContext context = getServiceContext();
-        DatacenterMO dcMo = null;
-        try {
-            VmwareManager mgr = context.getStockObject(VmwareManager.CONTEXT_STOCK_NAME);
-
-            VmwareHypervisorHost hyperHost = getHyperHost(context);
-            dcMo = new DatacenterMO(hyperHost.getContext(), hyperHost.getHyperHostDatacenter());
-
-            // Validate VM name is unique in Datacenter
-            VirtualMachineMO vmInVcenter = dcMo.checkIfVmAlreadyExistsInVcenter(vmNameOnVcenter, vmInternalCSName);
-            if (vmInVcenter != null) {
-                vmAlreadyExistsInVcenter = true;
-                String msg = "VM with name: " + vmNameOnVcenter + " already exists in vCenter.";
-                s_logger.error(msg);
-                throw new Exception(msg);
-            }
-            String guestOsId = translateGuestOsIdentifier(vmSpec.getArch(), vmSpec.getOs(), vmSpec.getPlatformEmulator()).value();
-            DiskTO[] disks = validateDisks(vmSpec.getDisks());
-            assert (disks.length > 0);
-            NicTO[] nics = vmSpec.getNics();
-
-            HashMap<String, Pair<ManagedObjectReference, DatastoreMO>> dataStoresDetails = inferDatastoreDetailsFromDiskInfo(hyperHost, context, disks, cmd);
-            if ((dataStoresDetails == null) || (dataStoresDetails.isEmpty())) {
-                String msg = "Unable to locate datastore details of the volumes to be attached";
-                s_logger.error(msg);
-                throw new Exception(msg);
-            }
-
-            DatastoreMO dsRootVolumeIsOn = getDatastoreThatRootDiskIsOn(dataStoresDetails, disks);
-            if (dsRootVolumeIsOn == null) {
-                String msg = "Unable to locate datastore details of root volume";
-                s_logger.error(msg);
-                throw new Exception(msg);
-            }
-
-            VirtualMachineDiskInfoBuilder diskInfoBuilder = null;
-            VirtualMachineMO vmMo = hyperHost.findVmOnHyperHost(vmInternalCSName);
-            DiskControllerType systemVmScsiControllerType = DiskControllerType.lsilogic;
-            int firstScsiControllerBusNum = 0;
-            int numScsiControllerForSystemVm = 1;
-            boolean hasSnapshot = false;
-            if (vmMo != null) {
-                s_logger.info("VM " + vmInternalCSName + " already exists, tear down devices for reconfiguration");
-                if (getVmPowerState(vmMo) != PowerState.PowerOff)
-                    vmMo.safePowerOff(_shutdownWaitMs);
-
-                // retrieve disk information before we tear down
-                diskInfoBuilder = vmMo.getDiskInfoBuilder();
-                hasSnapshot = vmMo.hasSnapshot();
-                if (!hasSnapshot)
-                    vmMo.tearDownDevices(new Class<?>[]{VirtualDisk.class, VirtualEthernetCard.class});
-                else
-                    vmMo.tearDownDevices(new Class<?>[]{VirtualEthernetCard.class});
-                if (systemVm) {
-                    ensureScsiDiskControllers(vmMo, systemVmScsiControllerType.toString(), numScsiControllerForSystemVm, firstScsiControllerBusNum);
-                } else {
-                    ensureDiskControllers(vmMo, controllerInfo);
-                }
-            } else {
-                ManagedObjectReference morDc = hyperHost.getHyperHostDatacenter();
-                assert (morDc != null);
-
-                vmMo = hyperHost.findVmOnPeerHyperHost(vmInternalCSName);
-                if (vmMo != null) {
-                    if (s_logger.isInfoEnabled()) {
-                        s_logger.info("Found vm " + vmInternalCSName + " at other host, relocate to " + hyperHost.getHyperHostName());
-                    }
-
-                    takeVmFromOtherHyperHost(hyperHost, vmInternalCSName);
-
-                    if (getVmPowerState(vmMo) != PowerState.PowerOff)
-                        vmMo.safePowerOff(_shutdownWaitMs);
-
-                    diskInfoBuilder = vmMo.getDiskInfoBuilder();
-                    hasSnapshot = vmMo.hasSnapshot();
-                    if (!hasSnapshot)
-                        vmMo.tearDownDevices(new Class<?>[]{VirtualDisk.class, VirtualEthernetCard.class});
-                    else
-                        vmMo.tearDownDevices(new Class<?>[]{VirtualEthernetCard.class});
-
-                    if (systemVm) {
-                        // System volumes doesn't require more than 1 SCSI controller as there is no requirement for data volumes.
-                        ensureScsiDiskControllers(vmMo, systemVmScsiControllerType.toString(), numScsiControllerForSystemVm, firstScsiControllerBusNum);
-                    } else {
-                        ensureDiskControllers(vmMo, controllerInfo);
-                    }
-                } else {
-                    // If a VM with the same name is found in a different cluster in the DC, unregister the old VM and configure a new VM (cold-migration).
-                    VirtualMachineMO existingVmInDc = dcMo.findVm(vmInternalCSName);
-                    if (existingVmInDc != null) {
-                        s_logger.debug("Found VM: " + vmInternalCSName + " on a host in a different cluster. Unregistering the exisitng VM.");
-                        existingVmName = existingVmInDc.getName();
-                        existingVmFileInfo = existingVmInDc.getFileInfo();
-                        existingVmFileLayout = existingVmInDc.getFileLayout();
-                        existingDatastores = existingVmInDc.getAllDatastores();
-                        existingVmInDc.unregisterVm();
-                    }
-                    Pair<ManagedObjectReference, DatastoreMO> rootDiskDataStoreDetails = null;
-                    for (DiskTO vol : disks) {
-                        if (vol.getType() == Volume.Type.ROOT) {
-                            Map<String, String> details = vol.getDetails();
-                            boolean managed = false;
-
-                            if (details != null) {
-                                managed = Boolean.parseBoolean(details.get(DiskTO.MANAGED));
-                            }
-
-                            if (managed) {
-                                String datastoreName = VmwareResource.getDatastoreName(details.get(DiskTO.IQN));
-
-                                rootDiskDataStoreDetails = dataStoresDetails.get(datastoreName);
-                            } else {
-                                DataStoreTO primaryStore = vol.getData().getDataStore();
-
-                                rootDiskDataStoreDetails = dataStoresDetails.get(primaryStore.getUuid());
-                            }
-                        }
-                    }
-
-                    assert (vmSpec.getMinSpeed() != null) && (rootDiskDataStoreDetails != null);
-
-                    boolean vmFolderExists = rootDiskDataStoreDetails.second().folderExists(String.format("[%s]", rootDiskDataStoreDetails.second().getName()), vmNameOnVcenter);
-                    String vmxFileFullPath = dsRootVolumeIsOn.searchFileInSubFolders(vmNameOnVcenter + ".vmx", false, VmwareManager.s_vmwareSearchExcludeFolder.value());
-                    if (vmFolderExists && vmxFileFullPath != null) { // VM can be registered only if .vmx is present.
-                        registerVm(vmNameOnVcenter, dsRootVolumeIsOn);
-                        vmMo = hyperHost.findVmOnHyperHost(vmInternalCSName);
-                        if (vmMo != null) {
-                            if (s_logger.isDebugEnabled()) {
-                                s_logger.debug("Found registered vm " + vmInternalCSName + " at host " + hyperHost.getHyperHostName());
-                            }
-                        }
-                        tearDownVm(vmMo);
-                    } else if (!hyperHost.createBlankVm(vmNameOnVcenter, vmInternalCSName, vmSpec.getCpus(), vmSpec.getMaxSpeed().intValue(), getReservedCpuMHZ(vmSpec),
-                            vmSpec.getLimitCpuUse(), (int) (vmSpec.getMaxRam() / ResourceType.bytesToMiB), getReservedMemoryMb(vmSpec), guestOsId, rootDiskDataStoreDetails.first(), false,
-                            controllerInfo, systemVm)) {
-                        throw new Exception("Failed to create VM. vmName: " + vmInternalCSName);
-                    }
-                }
-
-                vmMo = hyperHost.findVmOnHyperHost(vmInternalCSName);
-                if (vmMo == null) {
-                    throw new Exception("Failed to find the newly create or relocated VM. vmName: " + vmInternalCSName);
-                }
-            }
-
-            int totalChangeDevices = disks.length + nics.length;
-            // vApp cdrom device
-            // HACK ALERT: ovf properties might not be the only or defining feature of vApps; needs checking
-            if (s_logger.isTraceEnabled()) {
-                s_logger.trace("adding divice tie device count for vApp config ISO");
-            }
-            if (vmSpec.getOvfProperties() != null) {
-                totalChangeDevices++;
-            }
-
-            DiskTO volIso = null;
-            if (vmSpec.getType() != VirtualMachine.Type.User) {
-                // system VM needs a patch ISO
-                totalChangeDevices++;
-            } else {
-                volIso = getIsoDiskTO(disks);
-                if (volIso == null)
-                    totalChangeDevices++;
-            }
-
-            VirtualMachineConfigSpec vmConfigSpec = new VirtualMachineConfigSpec();
-
-            VmwareHelper.setBasicVmConfig(vmConfigSpec, vmSpec.getCpus(), vmSpec.getMaxSpeed(), getReservedCpuMHZ(vmSpec), (int) (vmSpec.getMaxRam() / (1024 * 1024)),
-                    getReservedMemoryMb(vmSpec), guestOsId, vmSpec.getLimitCpuUse());
-
-            // Check for multi-cores per socket settings
-            int numCoresPerSocket = 1;
-            String coresPerSocket = vmSpec.getDetails().get(VmDetailConstants.CPU_CORE_PER_SOCKET);
-            if (coresPerSocket != null) {
-                String apiVersion = HypervisorHostHelper.getVcenterApiVersion(vmMo.getContext());
-                // Property 'numCoresPerSocket' is supported since vSphere API 5.0
-                if (apiVersion.compareTo("5.0") >= 0) {
-                    numCoresPerSocket = NumbersUtil.parseInt(coresPerSocket, 1);
-                    vmConfigSpec.setNumCoresPerSocket(numCoresPerSocket);
-                }
-            }
-
-            // Check for hotadd settings
-            vmConfigSpec.setMemoryHotAddEnabled(vmMo.isMemoryHotAddSupported(guestOsId));
-
-            String hostApiVersion = ((HostMO) hyperHost).getHostAboutInfo().getApiVersion();
-            if (numCoresPerSocket > 1 && hostApiVersion.compareTo("5.0") < 0) {
-                s_logger.warn("Dynamic scaling of CPU is not supported for Virtual Machines with multi-core vCPUs in case of ESXi hosts 4.1 and prior. Hence CpuHotAdd will not be"
-                        + " enabled for Virtual Machine: " + vmInternalCSName);
-                vmConfigSpec.setCpuHotAddEnabled(false);
-            } else {
-                vmConfigSpec.setCpuHotAddEnabled(vmMo.isCpuHotAddSupported(guestOsId));
-            }
-
-            configNestedHVSupport(vmMo, vmSpec, vmConfigSpec);
-
-            VirtualDeviceConfigSpec[] deviceConfigSpecArray = new VirtualDeviceConfigSpec[totalChangeDevices];
-            int i = 0;
-            int ideUnitNumber = 0;
-            int scsiUnitNumber = 0;
-            int ideControllerKey = vmMo.getIDEDeviceControllerKey();
-            int scsiControllerKey = vmMo.getGenericScsiDeviceControllerKeyNoException();
-            int controllerKey;
-
-            //
-            // Setup ISO device
-            //
-
-            // vAPP ISO
-            // FR37 the native deploy mechs should create this for us
-            if (vmSpec.getOvfProperties() != null) {
-                if (s_logger.isTraceEnabled()) {
-                    // FR37 TODO add more usefull info (if we keep this bit
-                    s_logger.trace("adding iso for properties for 'xxx'");
-                }
-                deviceConfigSpecArray[i] = new VirtualDeviceConfigSpec();
-                Pair<VirtualDevice, Boolean> isoInfo = VmwareHelper.prepareIsoDevice(vmMo, null, null, true, true, ideUnitNumber++, i + 1);
-                deviceConfigSpecArray[i].setDevice(isoInfo.first());
-                if (isoInfo.second()) {
-                    if (s_logger.isDebugEnabled())
-                        s_logger.debug("Prepare vApp ISO volume at existing device " + _gson.toJson(isoInfo.first()));
-
-                    deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.ADD);
-                } else {
-                    if (s_logger.isDebugEnabled())
-                        s_logger.debug("Prepare vApp ISO volume at existing device " + _gson.toJson(isoInfo.first()));
-
-                    deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.EDIT);
-                }
-                i++;
-            }
-
-            // prepare systemvm patch ISO
-            if (vmSpec.getType() != VirtualMachine.Type.User) {
-                // attach ISO (for patching of system VM)
-                Pair<String, Long> secStoreUrlAndId = mgr.getSecondaryStorageStoreUrlAndId(Long.parseLong(_dcId));
-                String secStoreUrl = secStoreUrlAndId.first();
-                Long secStoreId = secStoreUrlAndId.second();
-                if (secStoreUrl == null) {
-                    String msg = "secondary storage for dc " + _dcId + " is not ready yet?";
-                    throw new Exception(msg);
-                }
-                mgr.prepareSecondaryStorageStore(secStoreUrl, secStoreId);
-
-                ManagedObjectReference morSecDs = prepareSecondaryDatastoreOnHost(secStoreUrl);
-                if (morSecDs == null) {
-                    String msg = "Failed to prepare secondary storage on host, secondary store url: " + secStoreUrl;
-                    throw new Exception(msg);
-                }
-                DatastoreMO secDsMo = new DatastoreMO(hyperHost.getContext(), morSecDs);
-
-                deviceConfigSpecArray[i] = new VirtualDeviceConfigSpec();
-                Pair<VirtualDevice, Boolean> isoInfo = VmwareHelper.prepareIsoDevice(vmMo,
-                        String.format("[%s] systemvm/%s", secDsMo.getName(), mgr.getSystemVMIsoFileNameOnDatastore()), secDsMo.getMor(), true, true, ideUnitNumber++, i + 1);
-                deviceConfigSpecArray[i].setDevice(isoInfo.first());
-                if (isoInfo.second()) {
-                    if (s_logger.isDebugEnabled())
-                        s_logger.debug("Prepare ISO volume at new device " + _gson.toJson(isoInfo.first()));
-                    deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.ADD);
-                } else {
-                    if (s_logger.isDebugEnabled())
-                        s_logger.debug("Prepare ISO volume at existing device " + _gson.toJson(isoInfo.first()));
-                    deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.EDIT);
-                }
-                i++;
-            } else {
-                // Note: we will always plug a CDROM device
-                if (volIso != null) {
-                    for (DiskTO vol : disks) {
-                        if (vol.getType() == Volume.Type.ISO) {
-
-                            TemplateObjectTO iso = (TemplateObjectTO) vol.getData();
-
-                            if (iso.getPath() != null && !iso.getPath().isEmpty()) {
-                                DataStoreTO imageStore = iso.getDataStore();
-                                if (!(imageStore instanceof NfsTO)) {
-                                    s_logger.debug("unsupported protocol");
-                                    throw new Exception("unsupported protocol");
-                                }
-                                NfsTO nfsImageStore = (NfsTO) imageStore;
-                                String isoPath = nfsImageStore.getUrl() + File.separator + iso.getPath();
-                                Pair<String, ManagedObjectReference> isoDatastoreInfo = getIsoDatastoreInfo(hyperHost, isoPath);
-                                assert (isoDatastoreInfo != null);
-                                assert (isoDatastoreInfo.second() != null);
-
-                                deviceConfigSpecArray[i] = new VirtualDeviceConfigSpec();
-                                Pair<VirtualDevice, Boolean> isoInfo =
-                                        VmwareHelper.prepareIsoDevice(vmMo, isoDatastoreInfo.first(), isoDatastoreInfo.second(), true, true, ideUnitNumber++, i + 1);
-                                deviceConfigSpecArray[i].setDevice(isoInfo.first());
-                                if (isoInfo.second()) {
-                                    if (s_logger.isDebugEnabled())
-                                        s_logger.debug("Prepare ISO volume at new device " + _gson.toJson(isoInfo.first()));
-                                    deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.ADD);
-                                } else {
-                                    if (s_logger.isDebugEnabled())
-                                        s_logger.debug("Prepare ISO volume at existing device " + _gson.toJson(isoInfo.first()));
-                                    deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.EDIT);
-                                }
-                            }
-                            i++;
-                        }
-                    }
-                } else {
-                    deviceConfigSpecArray[i] = new VirtualDeviceConfigSpec();
-                    Pair<VirtualDevice, Boolean> isoInfo = VmwareHelper.prepareIsoDevice(vmMo, null, null, true, true, ideUnitNumber++, i + 1);
-                    deviceConfigSpecArray[i].setDevice(isoInfo.first());
-                    if (isoInfo.second()) {
-                        if (s_logger.isDebugEnabled())
-                            s_logger.debug("Prepare ISO volume at existing device " + _gson.toJson(isoInfo.first()));
-
-                        deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.ADD);
-                    } else {
-                        if (s_logger.isDebugEnabled())
-                            s_logger.debug("Prepare ISO volume at existing device " + _gson.toJson(isoInfo.first()));
-
-                        deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.EDIT);
-                    }
-                    i++;
-                }
-            }
-
-
-            //
-            // Setup ROOT/DATA disk devices
-            //
-            DiskTO[] sortedDisks = sortVolumesByDeviceId(disks);
-            for (DiskTO vol : sortedDisks) {
-                if (vol.getType() == Volume.Type.ISO)
-                    continue;
-
-                VirtualMachineDiskInfo matchingExistingDisk = getMatchingExistingDisk(diskInfoBuilder, vol, hyperHost, context);
-                controllerKey = getDiskController(matchingExistingDisk, vol, vmSpec, ideControllerKey, scsiControllerKey);
-                String diskController = getDiskController(vmMo, matchingExistingDisk, vol, new Pair<String, String>(rootDiskController, dataDiskController));
-
-                if (DiskControllerType.getType(diskController) == DiskControllerType.osdefault) {
-                    diskController = vmMo.getRecommendedDiskController(null);
-                }
-                if (DiskControllerType.getType(diskController) == DiskControllerType.ide) {
-                    controllerKey = vmMo.getIDEControllerKey(ideUnitNumber);
-                    if (vol.getType() == Volume.Type.DATADISK) {
-                        // Could be result of flip due to user configured setting or "osdefault" for data disks
-                        // Ensure maximum of 2 data volumes over IDE controller, 3 includeing root volume
-                        if (vmMo.getNumberOfVirtualDisks() > 3) {
-                            throw new CloudRuntimeException("Found more than 3 virtual disks attached to this VM [" + vmMo.getVmName() + "]. Unable to implement the disks over "
-                                    + diskController + " controller, as maximum number of devices supported over IDE controller is 4 includeing CDROM device.");
-                        }
-                    }
-                } else {
-                    controllerKey = vmMo.getScsiDiskControllerKeyNoException(diskController);
-                    if (controllerKey == -1) {
-                        // This may happen for ROOT legacy VMs which doesn't have recommended disk controller when global configuration parameter 'vmware.root.disk.controller' is set to "osdefault"
-                        // Retrieve existing controller and use.
-                        Ternary<Integer, Integer, DiskControllerType> vmScsiControllerInfo = vmMo.getScsiControllerInfo();
-                        DiskControllerType existingControllerType = vmScsiControllerInfo.third();
-                        controllerKey = vmMo.getScsiDiskControllerKeyNoException(existingControllerType.toString());
-                    }
-                }
-                if (!hasSnapshot) {
-                    deviceConfigSpecArray[i] = new VirtualDeviceConfigSpec();
-
-                    VolumeObjectTO volumeTO = (VolumeObjectTO) vol.getData();
-                    DataStoreTO primaryStore = volumeTO.getDataStore();
-                    Map<String, String> details = vol.getDetails();
-                    boolean managed = false;
-                    String iScsiName = null;
-
-                    if (details != null) {
-                        managed = Boolean.parseBoolean(details.get(DiskTO.MANAGED));
-                        iScsiName = details.get(DiskTO.IQN);
-                    }
-
-                    // if the storage is managed, iScsiName should not be null
-                    String datastoreName = managed ? VmwareResource.getDatastoreName(iScsiName) : primaryStore.getUuid();
-                    Pair<ManagedObjectReference, DatastoreMO> volumeDsDetails = dataStoresDetails.get(datastoreName);
-
-                    assert (volumeDsDetails != null);
-
-                    String[] diskChain = syncDiskChain(dcMo, vmMo, vmSpec, vol, matchingExistingDisk, dataStoresDetails);
-                    if (controllerKey == scsiControllerKey && VmwareHelper.isReservedScsiDeviceNumber(scsiUnitNumber))
-                        scsiUnitNumber++;
-                    VirtualDevice device = VmwareHelper.prepareDiskDevice(vmMo, null, controllerKey, diskChain, volumeDsDetails.first(),
-                            (controllerKey == vmMo.getIDEControllerKey(ideUnitNumber)) ? ((ideUnitNumber++) % VmwareHelper.MAX_IDE_CONTROLLER_COUNT) : scsiUnitNumber++, i + 1);
-
-                    if (vol.getType() == Volume.Type.ROOT)
-                        rootDiskTO = vol;
-                    deviceConfigSpecArray[i].setDevice(device);
-                    deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.ADD);
-
-                    if (s_logger.isDebugEnabled())
-                        s_logger.debug("Prepare volume at new device " + _gson.toJson(device));
-
-                    i++;
-                } else {
-                    if (controllerKey == scsiControllerKey && VmwareHelper.isReservedScsiDeviceNumber(scsiUnitNumber))
-                        scsiUnitNumber++;
-                    if (controllerKey == vmMo.getIDEControllerKey(ideUnitNumber))
-                        ideUnitNumber++;
-                    else
-                        scsiUnitNumber++;
-                }
-            }
-
-            //
-            // Setup USB devices
-            //
-            if (guestOsId.startsWith("darwin")) { //Mac OS
-                VirtualDevice[] devices = vmMo.getMatchedDevices(new Class<?>[]{VirtualUSBController.class});
-                if (devices.length == 0) {
-                    s_logger.debug("No USB Controller device on VM Start. Add USB Controller device for Mac OS VM " + vmInternalCSName);
-
-                    //For Mac OS X systems, the EHCI+UHCI controller is enabled by default and is required for USB mouse and keyboard access.
-                    VirtualDevice usbControllerDevice = VmwareHelper.prepareUSBControllerDevice();
-                    deviceConfigSpecArray[i] = new VirtualDeviceConfigSpec();
-                    deviceConfigSpecArray[i].setDevice(usbControllerDevice);
-                    deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.ADD);
-
-                    if (s_logger.isDebugEnabled())
-                        s_logger.debug("Prepare USB controller at new device " + _gson.toJson(deviceConfigSpecArray[i]));
-
-                    i++;
-                } else {
-                    s_logger.debug("USB Controller device exists on VM Start for Mac OS VM " + vmInternalCSName);
-                }
-            }
-
-            //
-            // Setup NIC devices
-            //
-            VirtualDevice nic;
-            int nicMask = 0;
-            int nicCount = 0;
-
-            if (vmSpec.getType() == VirtualMachine.Type.DomainRouter) {
-                int extraPublicNics = mgr.getRouterExtraPublicNics();
-                if (extraPublicNics > 0 && vmSpec.getDetails().containsKey("PeerRouterInstanceName")) {
-                    //Set identical MAC address for RvR on extra public interfaces
-                    String peerRouterInstanceName = vmSpec.getDetails().get("PeerRouterInstanceName");
-
-                    VirtualMachineMO peerVmMo = hyperHost.findVmOnHyperHost(peerRouterInstanceName);
-                    if (peerVmMo == null) {
-                        peerVmMo = hyperHost.findVmOnPeerHyperHost(peerRouterInstanceName);
-                    }
-
-                    if (peerVmMo != null) {
-                        String oldMacSequence = generateMacSequence(nics);
-
-                        for (int nicIndex = nics.length - extraPublicNics; nicIndex < nics.length; nicIndex++) {
-                            VirtualDevice nicDevice = peerVmMo.getNicDeviceByIndex(nics[nicIndex].getDeviceId());
-                            if (nicDevice != null) {
-                                String mac = ((VirtualEthernetCard) nicDevice).getMacAddress();
-                                if (mac != null) {
-                                    s_logger.info("Use same MAC as previous RvR, the MAC is " + mac + " for extra NIC with device id: " + nics[nicIndex].getDeviceId());
-                                    nics[nicIndex].setMac(mac);
-                                }
-                            }
-                        }
-
-                        if (!StringUtils.isBlank(vmSpec.getBootArgs())) {
-                            String newMacSequence = generateMacSequence(nics);
-                            vmSpec.setBootArgs(replaceNicsMacSequenceInBootArgs(oldMacSequence, newMacSequence, vmSpec));
-                        }
-                    }
-                }
-            }
-
-            VirtualEthernetCardType nicDeviceType = VirtualEthernetCardType.valueOf(vmSpec.getDetails().get(VmDetailConstants.NIC_ADAPTER));
-            if (s_logger.isDebugEnabled())
-                s_logger.debug("VM " + vmInternalCSName + " will be started with NIC device type: " + nicDeviceType);
-
-            NiciraNvpApiVersion.logNiciraApiVersion();
-
-            Map<String, String> nicUuidToDvSwitchUuid = new HashMap<String, String>();
-            for (NicTO nicTo : sortNicsByDeviceId(nics)) {
-                s_logger.info("Prepare NIC device based on NicTO: " + _gson.toJson(nicTo));
-
-                boolean configureVServiceInNexus = (nicTo.getType() == TrafficType.Guest) && (vmSpec.getDetails().containsKey("ConfigureVServiceInNexus"));
-                VirtualMachine.Type vmType = cmd.getVirtualMachine().getType();
-                Pair<ManagedObjectReference, String> networkInfo = prepareNetworkFromNicInfo(vmMo.getRunningHost(), nicTo, configureVServiceInNexus, vmType);
-                if ((nicTo.getBroadcastType() != BroadcastDomainType.Lswitch)
-                        || (nicTo.getBroadcastType() == BroadcastDomainType.Lswitch && NiciraNvpApiVersion.isApiVersionLowerThan("4.2"))) {
-                    if (VmwareHelper.isDvPortGroup(networkInfo.first())) {
-                        String dvSwitchUuid;
-                        ManagedObjectReference dcMor = hyperHost.getHyperHostDatacenter();
-                        DatacenterMO dataCenterMo = new DatacenterMO(context, dcMor);
-                        ManagedObjectReference dvsMor = dataCenterMo.getDvSwitchMor(networkInfo.first());
-                        dvSwitchUuid = dataCenterMo.getDvSwitchUuid(dvsMor);
-                        s_logger.info("Preparing NIC device on dvSwitch : " + dvSwitchUuid);
-                        nic = VmwareHelper.prepareDvNicDevice(vmMo, networkInfo.first(), nicDeviceType, networkInfo.second(), dvSwitchUuid,
-                                nicTo.getMac(), i + 1, true, true);
-                        if (nicTo.getUuid() != null) {
-                            nicUuidToDvSwitchUuid.put(nicTo.getUuid(), dvSwitchUuid);
-                        }
-                    } else {
-                        s_logger.info("Preparing NIC device on network " + networkInfo.second());
-                        nic = VmwareHelper.prepareNicDevice(vmMo, networkInfo.first(), nicDeviceType, networkInfo.second(),
-                                nicTo.getMac(), i + 1, true, true);
-                    }
-                } else {
-                    //if NSX API VERSION >= 4.2, connect to br-int (nsx.network), do not create portgroup else previous behaviour
-                    nic = VmwareHelper.prepareNicOpaque(vmMo, nicDeviceType, networkInfo.second(),
-                            nicTo.getMac(), i + 1, true, true);
-                }
-
-                deviceConfigSpecArray[i] = new VirtualDeviceConfigSpec();
-                deviceConfigSpecArray[i].setDevice(nic);
-                deviceConfigSpecArray[i].setOperation(VirtualDeviceConfigSpecOperation.ADD);
-
-                if (s_logger.isDebugEnabled())
-                    s_logger.debug("Prepare NIC at new device " + _gson.toJson(deviceConfigSpecArray[i]));
-
-                // this is really a hacking for DomR, upon DomR startup, we will reset all the NIC allocation after eth3
-                if (nicCount < 3)
-                    nicMask |= (1 << nicCount);
-
-                i++;
-                nicCount++;
-            }
-
-            for (int j = 0; j < i; j++)
-                vmConfigSpec.getDeviceChange().add(deviceConfigSpecArray[j]);
-
-            //
-            // Setup VM options
-            //
-
-            // pass boot arguments through machine.id & perform customized options to VMX
-            ArrayList<OptionValue> extraOptions = new ArrayList<OptionValue>();
-            configBasicExtraOption(extraOptions, vmSpec);
-            configNvpExtraOption(extraOptions, vmSpec, nicUuidToDvSwitchUuid);
-            configCustomExtraOption(extraOptions, vmSpec);
-
-            // config for NCC
-            VirtualMachine.Type vmType = cmd.getVirtualMachine().getType();
-            if (vmType.equals(VirtualMachine.Type.NetScalerVm)) {
-                NicTO mgmtNic = vmSpec.getNics()[0];
-                OptionValue option = new OptionValue();
-                option.setKey("machine.id");
-                option.setValue("ip=" + mgmtNic.getIp() + "&netmask=" + mgmtNic.getNetmask() + "&gateway=" + mgmtNic.getGateway());
-                extraOptions.add(option);
-            }
-
-            // config VNC
-            String keyboardLayout = null;
-            if (vmSpec.getDetails() != null)
-                keyboardLayout = vmSpec.getDetails().get(VmDetailConstants.KEYBOARD);
-            vmConfigSpec.getExtraConfig()
-                    .addAll(Arrays.asList(configureVnc(extraOptions.toArray(new OptionValue[0]), hyperHost, vmInternalCSName, vmSpec.getVncPassword(), keyboardLayout)));
-
-            // config video card
-            configureVideoCard(vmMo, vmSpec, vmConfigSpec);
-
-            // Set OVF properties (if available)
-            Pair<String, List<OVFPropertyTO>> ovfPropsMap = vmSpec.getOvfProperties();
-            VmConfigInfo templateVappConfig = null;
-            List<OVFPropertyTO> ovfProperties = null;
-            if (ovfPropsMap != null) {
-                String vmTemplate = ovfPropsMap.first();
-                s_logger.info("Find VM template " + vmTemplate);
-                VirtualMachineMO vmTemplateMO = dcMo.findVm(vmTemplate);
-                templateVappConfig = vmTemplateMO.getConfigInfo().getVAppConfig();
-                ovfProperties = ovfPropsMap.second();
-                // Set OVF properties (if available)
-                if (CollectionUtils.isNotEmpty(ovfProperties)) {
-                    s_logger.info("Copying OVF properties from template and setting them to the values the user provided");
-                    copyVAppConfigsFromTemplate(templateVappConfig, ovfProperties, vmConfigSpec);
-                }
-            }
-
-            setBootOptions(vmSpec, bootMode, vmConfigSpec);
-
-            //
-            // Configure VM
-            //
-            if (!vmMo.configureVm(vmConfigSpec)) {
-                throw new Exception("Failed to configure VM before start. vmName: " + vmInternalCSName);
-            }
-
-            if (vmSpec.getType() == VirtualMachine.Type.DomainRouter) {
-                hyperHost.setRestartPriorityForVM(vmMo, DasVmPriority.HIGH.value());
-            }
-
-            // Resizing root disk only when explicit requested by user
-            final Map<String, String> vmDetails = cmd.getVirtualMachine().getDetails();
-            if (rootDiskTO != null && !hasSnapshot && (vmDetails != null && vmDetails.containsKey(ApiConstants.ROOT_DISK_SIZE))) {
-                resizeRootDiskOnVMStart(vmMo, rootDiskTO, hyperHost, context);
-            }
-
-            //
-            // Post Configuration
-            //
-
-            vmMo.setCustomFieldValue(CustomFieldConstants.CLOUD_NIC_MASK, String.valueOf(nicMask));
-            postNvpConfigBeforeStart(vmMo, vmSpec);
-
-            Map<String, Map<String, String>> iqnToData = new HashMap<>();
-
-            postDiskConfigBeforeStart(vmMo, vmSpec, sortedDisks, ideControllerKey, scsiControllerKey, iqnToData, hyperHost, context);
-
-            //
-            // Power-on VM
-            //
-            if (!vmMo.powerOn()) {
-                throw new Exception("Failed to start VM. vmName: " + vmInternalCSName + " with hostname " + vmNameOnVcenter);
-            }
-
-            StartAnswer startAnswer = new StartAnswer(cmd);
-
-            startAnswer.setIqnToData(iqnToData);
-
-            // Since VM was successfully powered-on, if there was an existing VM in a different cluster that was unregistered, delete all the files associated with it.
-            if (existingVmName != null && existingVmFileLayout != null) {
-                List<String> vmDatastoreNames = new ArrayList<String>();
-                for (DatastoreMO vmDatastore : vmMo.getAllDatastores()) {
-                    vmDatastoreNames.add(vmDatastore.getName());
-                }
-                // Don't delete files that are in a datastore that is being used by the new VM as well (zone-wide datastore).
-                List<String> skipDatastores = new ArrayList<String>();
-                for (DatastoreMO existingDatastore : existingDatastores) {
-                    if (vmDatastoreNames.contains(existingDatastore.getName())) {
-                        skipDatastores.add(existingDatastore.getName());
-                    }
-                }
-                deleteUnregisteredVmFiles(existingVmFileLayout, dcMo, true, skipDatastores);
-            }
-
-            return startAnswer;
-        } catch (Throwable e) {
-            if (e instanceof RemoteException) {
-                s_logger.warn("Encounter remote exception to vCenter, invalidate VMware session context");
-                invalidateServiceContext();
-            }
-
-            String msg = "StartCommand failed due to " + VmwareHelper.getExceptionMessage(e);
-            s_logger.warn(msg, e);
-            StartAnswer startAnswer = new StartAnswer(cmd, msg);
-            if (vmAlreadyExistsInVcenter) {
-                startAnswer.setContextParam("stopRetry", "true");
-            }
-
-            // Since VM start failed, if there was an existing VM in a different cluster that was unregistered, register it back.
-            if (existingVmName != null && existingVmFileInfo != null) {
-                s_logger.debug("Since VM start failed, registering back an existing VM: " + existingVmName + " that was unregistered");
-                try {
-                    DatastoreFile fileInDatastore = new DatastoreFile(existingVmFileInfo.getVmPathName());
-                    DatastoreMO existingVmDsMo = new DatastoreMO(dcMo.getContext(), dcMo.findDatastore(fileInDatastore.getDatastoreName()));
-                    registerVm(existingVmName, existingVmDsMo);
-                } catch (Exception ex) {
-                    String message = "Failed to register an existing VM: " + existingVmName + " due to " + VmwareHelper.getExceptionMessage(ex);
-                    s_logger.warn(message, ex);
-                }
-            }
-
-            return startAnswer;
-        } finally {
-        }
-    }
-
-    private void setBootOptions(VirtualMachineTO vmSpec, String bootMode, VirtualMachineConfigSpec vmConfigSpec) {
-        VirtualMachineBootOptions bootOptions = null;
-        if (StringUtils.isNotBlank(bootMode) && !bootMode.equalsIgnoreCase("bios")) {
-            vmConfigSpec.setFirmware("efi");
-            if (vmSpec.getDetails().containsKey(ApiConstants.BootType.UEFI.toString()) && "secure".equalsIgnoreCase(vmSpec.getDetails().get(ApiConstants.BootType.UEFI.toString()))) {
-                if (bootOptions == null) {
-                    bootOptions = new VirtualMachineBootOptions();
-                }
-                bootOptions.setEfiSecureBootEnabled(true);
-            }
-        }
-        if (vmSpec.isEnterHardwareSetup()) {
-            if (bootOptions == null) {
-                bootOptions = new VirtualMachineBootOptions();
-            }
-            if (s_logger.isDebugEnabled()) {
-                s_logger.debug(String.format("configuring VM '%s' to enter hardware setup",vmSpec.getName()));
-            }
-            bootOptions.setEnterBIOSSetup(vmSpec.isEnterHardwareSetup());
-        }
-        if (bootOptions != null) {
-            vmConfigSpec.setBootOptions(bootOptions);
-        }
+        return startCommandExecutor.execute(cmd);
     }
 
     /**
@@ -2521,74 +1681,12 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         vmConfig.setVAppConfig(vmConfigSpec);
     }
 
-    private String appendFileType(String path, String fileType) {
+    static String appendFileType(String path, String fileType) {
         if (path.toLowerCase().endsWith(fileType.toLowerCase())) {
             return path;
         }
 
         return path + fileType;
-    }
-
-    private void resizeRootDiskOnVMStart(VirtualMachineMO vmMo, DiskTO rootDiskTO, VmwareHypervisorHost hyperHost, VmwareContext context) throws Exception {
-        final Pair<VirtualDisk, String> vdisk = getVirtualDiskInfo(vmMo, appendFileType(rootDiskTO.getPath(), VMDK_EXTENSION));
-        assert (vdisk != null);
-
-        Long reqSize = 0L;
-        final VolumeObjectTO volumeTO = ((VolumeObjectTO) rootDiskTO.getData());
-        if (volumeTO != null) {
-            reqSize = volumeTO.getSize() / 1024;
-        }
-        final VirtualDisk disk = vdisk.first();
-        if (reqSize > disk.getCapacityInKB()) {
-            final VirtualMachineDiskInfo diskInfo = getMatchingExistingDisk(vmMo.getDiskInfoBuilder(), rootDiskTO, hyperHost, context);
-            assert (diskInfo != null);
-            final String[] diskChain = diskInfo.getDiskChain();
-
-            if (diskChain != null && diskChain.length > 1) {
-                s_logger.warn("Disk chain length for the VM is greater than one, this is not supported");
-                throw new CloudRuntimeException("Unsupported VM disk chain length: " + diskChain.length);
-            }
-
-            boolean resizingSupported = false;
-            String deviceBusName = diskInfo.getDiskDeviceBusName();
-            if (deviceBusName != null && (deviceBusName.toLowerCase().contains("scsi") || deviceBusName.toLowerCase().contains("lsi"))) {
-                resizingSupported = true;
-            }
-            if (!resizingSupported) {
-                s_logger.warn("Resizing of root disk is only support for scsi device/bus, the provide VM's disk device bus name is " + diskInfo.getDiskDeviceBusName());
-                throw new CloudRuntimeException("Unsupported VM root disk device bus: " + diskInfo.getDiskDeviceBusName());
-            }
-
-            disk.setCapacityInKB(reqSize);
-            VirtualMachineConfigSpec vmConfigSpec = new VirtualMachineConfigSpec();
-            VirtualDeviceConfigSpec deviceConfigSpec = new VirtualDeviceConfigSpec();
-            deviceConfigSpec.setDevice(disk);
-            deviceConfigSpec.setOperation(VirtualDeviceConfigSpecOperation.EDIT);
-            vmConfigSpec.getDeviceChange().add(deviceConfigSpec);
-            if (!vmMo.configureVm(vmConfigSpec)) {
-                throw new Exception("Failed to configure VM for given root disk size. vmName: " + vmMo.getName());
-            }
-        }
-    }
-
-
-    /**
-     * Generate the mac sequence from the nics.
-     */
-    protected String generateMacSequence(NicTO[] nics) {
-        if (nics.length == 0) {
-            return "";
-        }
-
-        StringBuffer sbMacSequence = new StringBuffer();
-        for (NicTO nicTo : sortNicsByDeviceId(nics)) {
-            sbMacSequence.append(nicTo.getMac()).append("|");
-        }
-        if (!sbMacSequence.toString().isEmpty()) {
-            sbMacSequence.deleteCharAt(sbMacSequence.length() - 1); //Remove extra '|' char appended at the end
-        }
-
-        return sbMacSequence.toString();
     }
 
     /**
@@ -2673,20 +1771,6 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         vmConfigSpec.getDeviceChange().add(arrayVideoCardConfigSpecs);
     }
 
-    private void tearDownVm(VirtualMachineMO vmMo) throws Exception {
-
-        if (vmMo == null)
-            return;
-
-        boolean hasSnapshot = false;
-        hasSnapshot = vmMo.hasSnapshot();
-        if (!hasSnapshot)
-            vmMo.tearDownDevices(new Class<?>[]{VirtualDisk.class, VirtualEthernetCard.class});
-        else
-            vmMo.tearDownDevices(new Class<?>[]{VirtualEthernetCard.class});
-        vmMo.ensureScsiDeviceController();
-    }
-
     int getReservedMemoryMb(VirtualMachineTO vmSpec) {
         if (vmSpec.getDetails().get(VMwareGuru.VmwareReserveMemory.key()).equalsIgnoreCase("true")) {
             return (int) (vmSpec.getMinRam() / ResourceType.bytesToMiB);
@@ -2699,84 +1783,6 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
             return vmSpec.getMinSpeed() * vmSpec.getCpus();
         }
         return 0;
-    }
-
-    // return the finalized disk chain for startup, from top to bottom
-    private String[] syncDiskChain(DatacenterMO dcMo, VirtualMachineMO vmMo, VirtualMachineTO vmSpec, DiskTO vol, VirtualMachineDiskInfo diskInfo,
-                                   HashMap<String, Pair<ManagedObjectReference, DatastoreMO>> dataStoresDetails) throws Exception {
-
-        VolumeObjectTO volumeTO = (VolumeObjectTO) vol.getData();
-        DataStoreTO primaryStore = volumeTO.getDataStore();
-        Map<String, String> details = vol.getDetails();
-        boolean isManaged = false;
-        String iScsiName = null;
-
-        if (details != null) {
-            isManaged = Boolean.parseBoolean(details.get(DiskTO.MANAGED));
-            iScsiName = details.get(DiskTO.IQN);
-        }
-
-        // if the storage is managed, iScsiName should not be null
-        String datastoreName = isManaged ? VmwareResource.getDatastoreName(iScsiName) : primaryStore.getUuid();
-        Pair<ManagedObjectReference, DatastoreMO> volumeDsDetails = dataStoresDetails.get(datastoreName);
-
-        if (volumeDsDetails == null) {
-            throw new Exception("Primary datastore " + primaryStore.getUuid() + " is not mounted on host.");
-        }
-
-        DatastoreMO dsMo = volumeDsDetails.second();
-
-        // we will honor vCenter's meta if it exists
-        if (diskInfo != null) {
-            // to deal with run-time upgrade to maintain the new datastore folder structure
-            String disks[] = diskInfo.getDiskChain();
-            for (int i = 0; i < disks.length; i++) {
-                DatastoreFile file = new DatastoreFile(disks[i]);
-                if (!isManaged && file.getDir() != null && file.getDir().isEmpty()) {
-                    s_logger.info("Perform run-time datastore folder upgrade. sync " + disks[i] + " to VM folder");
-                    disks[i] = VmwareStorageLayoutHelper.syncVolumeToVmDefaultFolder(dcMo, vmMo.getName(), dsMo, file.getFileBaseName(), VmwareManager.s_vmwareSearchExcludeFolder.value());
-                }
-            }
-            return disks;
-        }
-
-        final String datastoreDiskPath;
-
-        if (isManaged) {
-            String vmdkPath = new DatastoreFile(volumeTO.getPath()).getFileBaseName();
-
-            if (volumeTO.getVolumeType() == Volume.Type.ROOT) {
-                if (vmdkPath == null) {
-                    vmdkPath = volumeTO.getName();
-                }
-
-                datastoreDiskPath = VmwareStorageLayoutHelper.syncVolumeToVmDefaultFolder(dcMo, vmMo.getName(), dsMo, vmdkPath);
-            } else {
-                if (vmdkPath == null) {
-                    vmdkPath = dsMo.getName();
-                }
-
-                datastoreDiskPath = dsMo.getDatastorePath(vmdkPath + VMDK_EXTENSION);
-            }
-        } else {
-            datastoreDiskPath = VmwareStorageLayoutHelper.syncVolumeToVmDefaultFolder(dcMo, vmMo.getName(), dsMo, volumeTO.getPath(), VmwareManager.s_vmwareSearchExcludeFolder.value());
-        }
-
-        if (!dsMo.fileExists(datastoreDiskPath)) {
-            s_logger.warn("Volume " + volumeTO.getId() + " does not seem to exist on datastore, out of sync? path: " + datastoreDiskPath);
-        }
-
-        return new String[]{datastoreDiskPath};
-    }
-
-    // Pair<internal CS name, vCenter display name>
-    private Pair<String, String> composeVmNames(VirtualMachineTO vmSpec) {
-        String vmInternalCSName = vmSpec.getName();
-        String vmNameOnVcenter = vmSpec.getName();
-        if (_instanceNameFlag && vmSpec.getHostName() != null) {
-            vmNameOnVcenter = vmSpec.getHostName();
-        }
-        return new Pair<String, String>(vmInternalCSName, vmNameOnVcenter);
     }
 
     protected void configNestedHVSupport(VirtualMachineMO vmMo, VirtualMachineTO vmSpec, VirtualMachineConfigSpec vmConfigSpec) throws Exception {
@@ -2804,357 +1810,6 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         }
     }
 
-    private static void configBasicExtraOption(List<OptionValue> extraOptions, VirtualMachineTO vmSpec) {
-        OptionValue newVal = new OptionValue();
-        newVal.setKey("machine.id");
-        newVal.setValue(vmSpec.getBootArgs());
-        extraOptions.add(newVal);
-
-        newVal = new OptionValue();
-        newVal.setKey("devices.hotplug");
-        newVal.setValue("true");
-        extraOptions.add(newVal);
-    }
-
-    private static void configNvpExtraOption(List<OptionValue> extraOptions, VirtualMachineTO vmSpec, Map<String, String> nicUuidToDvSwitchUuid) {
-        /**
-         * Extra Config : nvp.vm-uuid = uuid
-         *  - Required for Nicira NVP integration
-         */
-        OptionValue newVal = new OptionValue();
-        newVal.setKey("nvp.vm-uuid");
-        newVal.setValue(vmSpec.getUuid());
-        extraOptions.add(newVal);
-
-        /**
-         * Extra Config : nvp.iface-id.<num> = uuid
-         *  - Required for Nicira NVP integration
-         */
-        int nicNum = 0;
-        for (NicTO nicTo : sortNicsByDeviceId(vmSpec.getNics())) {
-            if (nicTo.getUuid() != null) {
-                newVal = new OptionValue();
-                newVal.setKey("nvp.iface-id." + nicNum);
-                newVal.setValue(nicTo.getUuid());
-                extraOptions.add(newVal);
-            }
-            nicNum++;
-        }
-    }
-
-    private static void configCustomExtraOption(List<OptionValue> extraOptions, VirtualMachineTO vmSpec) {
-        // we no longer to validation anymore
-        for (Map.Entry<String, String> entry : vmSpec.getDetails().entrySet()) {
-            if (entry.getKey().equalsIgnoreCase(VmDetailConstants.BOOT_MODE)) {
-                continue;
-            }
-            OptionValue newVal = new OptionValue();
-            newVal.setKey(entry.getKey());
-            newVal.setValue(entry.getValue());
-            extraOptions.add(newVal);
-        }
-    }
-
-    private static void postNvpConfigBeforeStart(VirtualMachineMO vmMo, VirtualMachineTO vmSpec) throws Exception {
-        /**
-         * We need to configure the port on the DV switch after the host is
-         * connected. So make this happen between the configure and start of
-         * the VM
-         */
-        int nicIndex = 0;
-        for (NicTO nicTo : sortNicsByDeviceId(vmSpec.getNics())) {
-            if (nicTo.getBroadcastType() == BroadcastDomainType.Lswitch) {
-                // We need to create a port with a unique vlan and pass the key to the nic device
-                s_logger.trace("Nic " + nicTo.toString() + " is connected to an NVP logicalswitch");
-                VirtualDevice nicVirtualDevice = vmMo.getNicDeviceByIndex(nicIndex);
-                if (nicVirtualDevice == null) {
-                    throw new Exception("Failed to find a VirtualDevice for nic " + nicIndex); //FIXME Generic exceptions are bad
-                }
-                VirtualDeviceBackingInfo backing = nicVirtualDevice.getBacking();
-                if (backing instanceof VirtualEthernetCardDistributedVirtualPortBackingInfo) {
-                    // This NIC is connected to a Distributed Virtual Switch
-                    VirtualEthernetCardDistributedVirtualPortBackingInfo portInfo = (VirtualEthernetCardDistributedVirtualPortBackingInfo) backing;
-                    DistributedVirtualSwitchPortConnection port = portInfo.getPort();
-                    String portKey = port.getPortKey();
-                    String portGroupKey = port.getPortgroupKey();
-                    String dvSwitchUuid = port.getSwitchUuid();
-
-                    s_logger.debug("NIC " + nicTo.toString() + " is connected to dvSwitch " + dvSwitchUuid + " pg " + portGroupKey + " port " + portKey);
-
-                    ManagedObjectReference dvSwitchManager = vmMo.getContext().getVimClient().getServiceContent().getDvSwitchManager();
-                    ManagedObjectReference dvSwitch = vmMo.getContext().getVimClient().getService().queryDvsByUuid(dvSwitchManager, dvSwitchUuid);
-
-                    // Get all ports
-                    DistributedVirtualSwitchPortCriteria criteria = new DistributedVirtualSwitchPortCriteria();
-                    criteria.setInside(true);
-                    criteria.getPortgroupKey().add(portGroupKey);
-                    List<DistributedVirtualPort> dvPorts = vmMo.getContext().getVimClient().getService().fetchDVPorts(dvSwitch, criteria);
-
-                    DistributedVirtualPort vmDvPort = null;
-                    List<Integer> usedVlans = new ArrayList<Integer>();
-                    for (DistributedVirtualPort dvPort : dvPorts) {
-                        // Find the port for this NIC by portkey
-                        if (portKey.equals(dvPort.getKey())) {
-                            vmDvPort = dvPort;
-                        }
-                        VMwareDVSPortSetting settings = (VMwareDVSPortSetting) dvPort.getConfig().getSetting();
-                        VmwareDistributedVirtualSwitchVlanIdSpec vlanId = (VmwareDistributedVirtualSwitchVlanIdSpec) settings.getVlan();
-                        s_logger.trace("Found port " + dvPort.getKey() + " with vlan " + vlanId.getVlanId());
-                        if (vlanId.getVlanId() > 0 && vlanId.getVlanId() < 4095) {
-                            usedVlans.add(vlanId.getVlanId());
-                        }
-                    }
-
-                    if (vmDvPort == null) {
-                        throw new Exception("Empty port list from dvSwitch for nic " + nicTo.toString());
-                    }
-
-                    DVPortConfigInfo dvPortConfigInfo = vmDvPort.getConfig();
-                    VMwareDVSPortSetting settings = (VMwareDVSPortSetting) dvPortConfigInfo.getSetting();
-
-                    VmwareDistributedVirtualSwitchVlanIdSpec vlanId = (VmwareDistributedVirtualSwitchVlanIdSpec) settings.getVlan();
-                    BoolPolicy blocked = settings.getBlocked();
-                    if (blocked.isValue() == Boolean.TRUE) {
-                        s_logger.trace("Port is blocked, set a vlanid and unblock");
-                        DVPortConfigSpec dvPortConfigSpec = new DVPortConfigSpec();
-                        VMwareDVSPortSetting edittedSettings = new VMwareDVSPortSetting();
-                        // Unblock
-                        blocked.setValue(Boolean.FALSE);
-                        blocked.setInherited(Boolean.FALSE);
-                        edittedSettings.setBlocked(blocked);
-                        // Set vlan
-                        int i;
-                        for (i = 1; i < 4095; i++) {
-                            if (!usedVlans.contains(i))
-                                break;
-                        }
-                        vlanId.setVlanId(i); // FIXME should be a determined
-                        // based on usage
-                        vlanId.setInherited(false);
-                        edittedSettings.setVlan(vlanId);
-
-                        dvPortConfigSpec.setSetting(edittedSettings);
-                        dvPortConfigSpec.setOperation("edit");
-                        dvPortConfigSpec.setKey(portKey);
-                        List<DVPortConfigSpec> dvPortConfigSpecs = new ArrayList<DVPortConfigSpec>();
-                        dvPortConfigSpecs.add(dvPortConfigSpec);
-                        ManagedObjectReference task = vmMo.getContext().getVimClient().getService().reconfigureDVPortTask(dvSwitch, dvPortConfigSpecs);
-                        if (!vmMo.getContext().getVimClient().waitForTask(task)) {
-                            throw new Exception("Failed to configure the dvSwitch port for nic " + nicTo.toString());
-                        }
-                        s_logger.debug("NIC " + nicTo.toString() + " connected to vlan " + i);
-                    } else {
-                        s_logger.trace("Port already configured and set to vlan " + vlanId.getVlanId());
-                    }
-                } else if (backing instanceof VirtualEthernetCardNetworkBackingInfo) {
-                    // This NIC is connected to a Virtual Switch
-                    // Nothing to do
-                } else if (backing instanceof VirtualEthernetCardOpaqueNetworkBackingInfo) {
-                    //if NSX API VERSION >= 4.2, connect to br-int (nsx.network), do not create portgroup else previous behaviour
-                    //OK, connected to OpaqueNetwork
-                } else {
-                    s_logger.error("nic device backing is of type " + backing.getClass().getName());
-                    throw new Exception("Incompatible backing for a VirtualDevice for nic " + nicIndex); //FIXME Generic exceptions are bad
-                }
-            }
-            nicIndex++;
-        }
-    }
-
-    private VirtualMachineDiskInfo getMatchingExistingDisk(VirtualMachineDiskInfoBuilder diskInfoBuilder, DiskTO vol, VmwareHypervisorHost hyperHost, VmwareContext context)
-            throws Exception {
-        if (diskInfoBuilder != null) {
-            VolumeObjectTO volume = (VolumeObjectTO) vol.getData();
-
-            String dsName = null;
-            String diskBackingFileBaseName = null;
-
-            Map<String, String> details = vol.getDetails();
-            boolean isManaged = details != null && Boolean.parseBoolean(details.get(DiskTO.MANAGED));
-
-            if (isManaged) {
-                String iScsiName = details.get(DiskTO.IQN);
-
-                // if the storage is managed, iScsiName should not be null
-                dsName = VmwareResource.getDatastoreName(iScsiName);
-
-                diskBackingFileBaseName = new DatastoreFile(volume.getPath()).getFileBaseName();
-            } else {
-                ManagedObjectReference morDs = HypervisorHostHelper.findDatastoreWithBackwardsCompatibility(hyperHost, volume.getDataStore().getUuid());
-                DatastoreMO dsMo = new DatastoreMO(context, morDs);
-
-                dsName = dsMo.getName();
-
-                diskBackingFileBaseName = volume.getPath();
-            }
-
-            VirtualMachineDiskInfo diskInfo = diskInfoBuilder.getDiskInfoByBackingFileBaseName(diskBackingFileBaseName, dsName);
-            if (diskInfo != null) {
-                s_logger.info("Found existing disk info from volume path: " + volume.getPath());
-                return diskInfo;
-            } else {
-                String chainInfo = volume.getChainInfo();
-                if (chainInfo != null) {
-                    VirtualMachineDiskInfo infoInChain = _gson.fromJson(chainInfo, VirtualMachineDiskInfo.class);
-                    if (infoInChain != null) {
-                        String[] disks = infoInChain.getDiskChain();
-                        if (disks.length > 0) {
-                            for (String diskPath : disks) {
-                                DatastoreFile file = new DatastoreFile(diskPath);
-                                diskInfo = diskInfoBuilder.getDiskInfoByBackingFileBaseName(file.getFileBaseName(), dsName);
-                                if (diskInfo != null) {
-                                    s_logger.info("Found existing disk from chain info: " + diskPath);
-                                    return diskInfo;
-                                }
-                            }
-                        }
-
-                        if (diskInfo == null) {
-                            diskInfo = diskInfoBuilder.getDiskInfoByDeviceBusName(infoInChain.getDiskDeviceBusName());
-                            if (diskInfo != null) {
-                                s_logger.info("Found existing disk from from chain device bus information: " + infoInChain.getDiskDeviceBusName());
-                                return diskInfo;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        return null;
-    }
-
-    private int getDiskController(VirtualMachineDiskInfo matchingExistingDisk, DiskTO vol, VirtualMachineTO vmSpec, int ideControllerKey, int scsiControllerKey) {
-
-        int controllerKey;
-        if (matchingExistingDisk != null) {
-            s_logger.info("Chose disk controller based on existing information: " + matchingExistingDisk.getDiskDeviceBusName());
-            if (matchingExistingDisk.getDiskDeviceBusName().startsWith("ide"))
-                return ideControllerKey;
-            else
-                return scsiControllerKey;
-        }
-
-        if (vol.getType() == Volume.Type.ROOT) {
-            Map<String, String> vmDetails = vmSpec.getDetails();
-            if (vmDetails != null && vmDetails.get(VmDetailConstants.ROOT_DISK_CONTROLLER) != null) {
-                if (vmDetails.get(VmDetailConstants.ROOT_DISK_CONTROLLER).equalsIgnoreCase("scsi")) {
-                    s_logger.info("Chose disk controller for vol " + vol.getType() + " -> scsi, based on root disk controller settings: "
-                            + vmDetails.get(VmDetailConstants.ROOT_DISK_CONTROLLER));
-                    controllerKey = scsiControllerKey;
-                } else {
-                    s_logger.info("Chose disk controller for vol " + vol.getType() + " -> ide, based on root disk controller settings: "
-                            + vmDetails.get(VmDetailConstants.ROOT_DISK_CONTROLLER));
-                    controllerKey = ideControllerKey;
-                }
-            } else {
-                s_logger.info("Chose disk controller for vol " + vol.getType() + " -> scsi. due to null root disk controller setting");
-                controllerKey = scsiControllerKey;
-            }
-
-        } else {
-            // DATA volume always use SCSI device
-            s_logger.info("Chose disk controller for vol " + vol.getType() + " -> scsi");
-            controllerKey = scsiControllerKey;
-        }
-
-        return controllerKey;
-    }
-
-    private String getDiskController(VirtualMachineMO vmMo, VirtualMachineDiskInfo matchingExistingDisk, DiskTO vol, Pair<String, String> controllerInfo) throws Exception {
-        int controllerKey;
-        DiskControllerType controllerType = DiskControllerType.none;
-        if (matchingExistingDisk != null) {
-            String currentBusName = matchingExistingDisk.getDiskDeviceBusName();
-            if (currentBusName != null) {
-                s_logger.info("Chose disk controller based on existing information: " + currentBusName);
-                if (currentBusName.startsWith("ide")) {
-                    controllerType = DiskControllerType.ide;
-                } else if (currentBusName.startsWith("scsi")) {
-                    controllerType = DiskControllerType.scsi;
-                }
-            }
-            if (controllerType == DiskControllerType.scsi || controllerType == DiskControllerType.none) {
-                Ternary<Integer, Integer, DiskControllerType> vmScsiControllerInfo = vmMo.getScsiControllerInfo();
-                controllerType = vmScsiControllerInfo.third();
-            }
-            return controllerType.toString();
-        }
-
-        if (vol.getType() == Volume.Type.ROOT) {
-            s_logger.info("Chose disk controller for vol " + vol.getType() + " -> " + controllerInfo.first()
-                    + ", based on root disk controller settings at global configuration setting.");
-            return controllerInfo.first();
-        } else {
-            s_logger.info("Chose disk controller for vol " + vol.getType() + " -> " + controllerInfo.second()
-                    + ", based on default data disk controller setting i.e. Operating system recommended."); // Need to bring in global configuration setting & template level setting.
-            return controllerInfo.second();
-        }
-    }
-
-    private void postDiskConfigBeforeStart(VirtualMachineMO vmMo, VirtualMachineTO vmSpec, DiskTO[] sortedDisks, int ideControllerKey,
-                                           int scsiControllerKey, Map<String, Map<String, String>> iqnToData, VmwareHypervisorHost hyperHost, VmwareContext context) throws Exception {
-        VirtualMachineDiskInfoBuilder diskInfoBuilder = vmMo.getDiskInfoBuilder();
-
-        for (DiskTO vol : sortedDisks) {
-            if (vol.getType() == Volume.Type.ISO)
-                continue;
-
-            VolumeObjectTO volumeTO = (VolumeObjectTO) vol.getData();
-
-            VirtualMachineDiskInfo diskInfo = getMatchingExistingDisk(diskInfoBuilder, vol, hyperHost, context);
-            assert (diskInfo != null);
-
-            String[] diskChain = diskInfo.getDiskChain();
-            assert (diskChain.length > 0);
-
-            Map<String, String> details = vol.getDetails();
-            boolean managed = false;
-
-            if (details != null) {
-                managed = Boolean.parseBoolean(details.get(DiskTO.MANAGED));
-            }
-
-            DatastoreFile file = new DatastoreFile(diskChain[0]);
-
-            if (managed) {
-                DatastoreFile originalFile = new DatastoreFile(volumeTO.getPath());
-
-                if (!file.getFileBaseName().equalsIgnoreCase(originalFile.getFileBaseName())) {
-                    if (s_logger.isInfoEnabled())
-                        s_logger.info("Detected disk-chain top file change on volume: " + volumeTO.getId() + " " + volumeTO.getPath() + " -> " + diskChain[0]);
-                }
-            } else {
-                if (!file.getFileBaseName().equalsIgnoreCase(volumeTO.getPath())) {
-                    if (s_logger.isInfoEnabled())
-                        s_logger.info("Detected disk-chain top file change on volume: " + volumeTO.getId() + " " + volumeTO.getPath() + " -> " + file.getFileBaseName());
-                }
-            }
-
-            VolumeObjectTO volInSpec = getVolumeInSpec(vmSpec, volumeTO);
-
-            if (volInSpec != null) {
-                if (managed) {
-                    Map<String, String> data = new HashMap<>();
-
-                    String datastoreVolumePath = diskChain[0];
-
-                    data.put(StartAnswer.PATH, datastoreVolumePath);
-                    data.put(StartAnswer.IMAGE_FORMAT, Storage.ImageFormat.OVA.toString());
-
-                    iqnToData.put(details.get(DiskTO.IQN), data);
-
-                    vol.setPath(datastoreVolumePath);
-                    volumeTO.setPath(datastoreVolumePath);
-                    volInSpec.setPath(datastoreVolumePath);
-                } else {
-                    volInSpec.setPath(file.getFileBaseName());
-                }
-                volInSpec.setChainInfo(_gson.toJson(diskInfo));
-            }
-        }
-    }
-
     private void checkAndDeleteDatastoreFile(String filePath, List<String> skipDatastores, DatastoreMO dsMo, DatacenterMO dcMo) throws Exception {
         if (dsMo != null && dcMo != null && (skipDatastores == null || !skipDatastores.contains(dsMo.getName()))) {
             s_logger.debug("Deleting file: " + filePath);
@@ -3162,7 +1817,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         }
     }
 
-    private void deleteUnregisteredVmFiles(VirtualMachineFileLayoutEx vmFileLayout, DatacenterMO dcMo, boolean deleteDisks, List<String> skipDatastores) throws Exception {
+    void deleteUnregisteredVmFiles(VirtualMachineFileLayoutEx vmFileLayout, DatacenterMO dcMo, boolean deleteDisks, List<String> skipDatastores) throws Exception {
         s_logger.debug("Deleting files associated with an existing VM that was unregistered");
         DatastoreFile vmFolder = null;
         try {
@@ -3196,64 +1851,6 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         }
     }
 
-    private static VolumeObjectTO getVolumeInSpec(VirtualMachineTO vmSpec, VolumeObjectTO srcVol) {
-        for (DiskTO disk : vmSpec.getDisks()) {
-            if (disk.getData() instanceof VolumeObjectTO) {
-                VolumeObjectTO vol = (VolumeObjectTO) disk.getData();
-                if (vol.getId() == srcVol.getId())
-                    return vol;
-            }
-        }
-
-        return null;
-    }
-
-    private static NicTO[] sortNicsByDeviceId(NicTO[] nics) {
-
-        List<NicTO> listForSort = new ArrayList<NicTO>();
-        for (NicTO nic : nics) {
-            listForSort.add(nic);
-        }
-        Collections.sort(listForSort, new Comparator<NicTO>() {
-
-            @Override
-            public int compare(NicTO arg0, NicTO arg1) {
-                if (arg0.getDeviceId() < arg1.getDeviceId()) {
-                    return -1;
-                } else if (arg0.getDeviceId() == arg1.getDeviceId()) {
-                    return 0;
-                }
-
-                return 1;
-            }
-        });
-
-        return listForSort.toArray(new NicTO[0]);
-    }
-
-    private static DiskTO[] sortVolumesByDeviceId(DiskTO[] volumes) {
-
-        List<DiskTO> listForSort = new ArrayList<DiskTO>();
-        for (DiskTO vol : volumes) {
-            listForSort.add(vol);
-        }
-        Collections.sort(listForSort, new Comparator<DiskTO>() {
-
-            @Override
-            public int compare(DiskTO arg0, DiskTO arg1) {
-                if (arg0.getDiskSeq() < arg1.getDiskSeq()) {
-                    return -1;
-                } else if (arg0.getDiskSeq().equals(arg1.getDiskSeq())) {
-                    return 0;
-                }
-
-                return 1;
-            }
-        });
-
-        return listForSort.toArray(new DiskTO[0]);
-    }
-
     /**
      * Only call this for managed storage.
      * Ex. "[-iqn.2010-01.com.solidfire:4nhe.vol-1.27-0] i-2-18-VM/ROOT-18.vmdk" should return "i-2-18-VM/ROOT-18"
@@ -3273,122 +1870,13 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
         path = path.substring(startIndex + search.length());
 
-        final String search2 = VMDK_EXTENSION;
-
-        int endIndex = path.indexOf(search2);
+        int endIndex = path.indexOf(VMDK_EXTENSION);
 
         if (endIndex == -1) {
             return null;
         }
 
         return path.substring(0, endIndex).trim();
-    }
-
-    private HashMap<String, Pair<ManagedObjectReference, DatastoreMO>> inferDatastoreDetailsFromDiskInfo(VmwareHypervisorHost hyperHost, VmwareContext context,
-                                                                                                         DiskTO[] disks, Command cmd) throws Exception {
-        HashMap<String, Pair<ManagedObjectReference, DatastoreMO>> mapIdToMors = new HashMap<>();
-
-        assert (hyperHost != null) && (context != null);
-
-        for (DiskTO vol : disks) {
-            if (vol.getType() != Volume.Type.ISO) {
-                VolumeObjectTO volumeTO = (VolumeObjectTO) vol.getData();
-                DataStoreTO primaryStore = volumeTO.getDataStore();
-                String poolUuid = primaryStore.getUuid();
-
-                if (mapIdToMors.get(poolUuid) == null) {
-                    boolean isManaged = false;
-                    Map<String, String> details = vol.getDetails();
-
-                    if (details != null) {
-                        isManaged = Boolean.parseBoolean(details.get(DiskTO.MANAGED));
-                    }
-
-                    if (isManaged) {
-                        String iScsiName = details.get(DiskTO.IQN); // details should not be null for managed storage (it may or may not be null for non-managed storage)
-                        String datastoreName = VmwareResource.getDatastoreName(iScsiName);
-                        ManagedObjectReference morDatastore = HypervisorHostHelper.findDatastoreWithBackwardsCompatibility(hyperHost, datastoreName);
-
-                        // if the datastore is not present, we need to discover the iSCSI device that will support it,
-                        // create the datastore, and create a VMDK file in the datastore
-                        if (morDatastore == null) {
-                            final String vmdkPath = getVmdkPath(volumeTO.getPath());
-
-                            morDatastore = _storageProcessor.prepareManagedStorage(context, hyperHost, null, iScsiName,
-                                    details.get(DiskTO.STORAGE_HOST), Integer.parseInt(details.get(DiskTO.STORAGE_PORT)),
-                                    vmdkPath,
-                                    details.get(DiskTO.CHAP_INITIATOR_USERNAME), details.get(DiskTO.CHAP_INITIATOR_SECRET),
-                                    details.get(DiskTO.CHAP_TARGET_USERNAME), details.get(DiskTO.CHAP_TARGET_SECRET),
-                                    Long.parseLong(details.get(DiskTO.VOLUME_SIZE)), cmd);
-
-                            DatastoreMO dsMo = new DatastoreMO(getServiceContext(), morDatastore);
-
-                            final String datastoreVolumePath;
-
-                            if (vmdkPath != null) {
-                                datastoreVolumePath = dsMo.getDatastorePath(vmdkPath + VMDK_EXTENSION);
-                            } else {
-                                datastoreVolumePath = dsMo.getDatastorePath(dsMo.getName() + VMDK_EXTENSION);
-                            }
-
-                            volumeTO.setPath(datastoreVolumePath);
-                            vol.setPath(datastoreVolumePath);
-                        }
-
-                        mapIdToMors.put(datastoreName, new Pair<>(morDatastore, new DatastoreMO(context, morDatastore)));
-                    } else {
-                        ManagedObjectReference morDatastore = HypervisorHostHelper.findDatastoreWithBackwardsCompatibility(hyperHost, poolUuid);
-
-                        if (morDatastore == null) {
-                            String msg = "Failed to get the mounted datastore for the volume's pool " + poolUuid;
-
-                            s_logger.error(msg);
-
-                            throw new Exception(msg);
-                        }
-
-                        mapIdToMors.put(poolUuid, new Pair<>(morDatastore, new DatastoreMO(context, morDatastore)));
-                    }
-                }
-            }
-        }
-
-        return mapIdToMors;
-    }
-
-    private DatastoreMO getDatastoreThatRootDiskIsOn(HashMap<String, Pair<ManagedObjectReference, DatastoreMO>> dataStoresDetails, DiskTO disks[]) {
-        Pair<ManagedObjectReference, DatastoreMO> rootDiskDataStoreDetails = null;
-
-        for (DiskTO vol : disks) {
-            if (vol.getType() == Volume.Type.ROOT) {
-                Map<String, String> details = vol.getDetails();
-                boolean managed = false;
-
-                if (details != null) {
-                    managed = Boolean.parseBoolean(details.get(DiskTO.MANAGED));
-                }
-
-                if (managed) {
-                    String datastoreName = VmwareResource.getDatastoreName(details.get(DiskTO.IQN));
-
-                    rootDiskDataStoreDetails = dataStoresDetails.get(datastoreName);
-
-                    break;
-                } else {
-                    DataStoreTO primaryStore = vol.getData().getDataStore();
-
-                    rootDiskDataStoreDetails = dataStoresDetails.get(primaryStore.getUuid());
-
-                    break;
-                }
-            }
-        }
-
-        if (rootDiskDataStoreDetails != null) {
-            return rootDiskDataStoreDetails.second();
-        }
-
-        return null;
     }
 
     private String getPvlanInfo(NicTO nicTo) {
@@ -3431,7 +1919,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         return defaultVlan;
     }
 
-    private Pair<ManagedObjectReference, String> prepareNetworkFromNicInfo(HostMO hostMo, NicTO nicTo, boolean configureVServiceInNexus, VirtualMachine.Type vmType)
+    Pair<ManagedObjectReference, String> prepareNetworkFromNicInfo(HostMO hostMo, NicTO nicTo, boolean configureVServiceInNexus, VirtualMachine.Type vmType)
             throws Exception {
 
         Ternary<String, String, String> switchDetails = getTargetSwitch(nicTo);
@@ -3532,58 +2020,9 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         }
     }
 
-    private VirtualMachineMO takeVmFromOtherHyperHost(VmwareHypervisorHost hyperHost, String vmName) throws Exception {
-
-        VirtualMachineMO vmMo = hyperHost.findVmOnPeerHyperHost(vmName);
-        if (vmMo != null) {
-            ManagedObjectReference morTargetPhysicalHost = hyperHost.findMigrationTarget(vmMo);
-            if (morTargetPhysicalHost == null) {
-                String msg = "VM " + vmName + " is on other host and we have no resource available to migrate and start it here";
-                s_logger.error(msg);
-                throw new Exception(msg);
-            }
-
-            if (!vmMo.relocate(morTargetPhysicalHost)) {
-                String msg = "VM " + vmName + " is on other host and we failed to relocate it here";
-                s_logger.error(msg);
-                throw new Exception(msg);
-            }
-
-            return vmMo;
-        }
-        return null;
-    }
-
-    // isoUrl sample content :
-    // nfs://192.168.10.231/export/home/kelven/vmware-test/secondary/template/tmpl/2/200//200-2-80f7ee58-6eff-3a2d-bcb0-59663edf6d26.iso
-    private Pair<String, ManagedObjectReference> getIsoDatastoreInfo(VmwareHypervisorHost hyperHost, String isoUrl) throws Exception {
-
-        assert (isoUrl != null);
-        int isoFileNameStartPos = isoUrl.lastIndexOf("/");
-        if (isoFileNameStartPos < 0) {
-            throw new Exception("Invalid ISO path info");
-        }
-
-        String isoFileName = isoUrl.substring(isoFileNameStartPos);
-
-        int templateRootPos = isoUrl.indexOf("template/tmpl");
-        templateRootPos = (templateRootPos < 0 ? isoUrl.indexOf(ConfigDrive.CONFIGDRIVEDIR) : templateRootPos);
-        if (templateRootPos < 0) {
-            throw new Exception("Invalid ISO path info");
-        }
-
-        String storeUrl = isoUrl.substring(0, templateRootPos - 1);
-        String isoPath = isoUrl.substring(templateRootPos, isoFileNameStartPos);
-
-        ManagedObjectReference morDs = prepareSecondaryDatastoreOnHost(storeUrl);
-        DatastoreMO dsMo = new DatastoreMO(getServiceContext(), morDs);
-
-        return new Pair<String, ManagedObjectReference>(String.format("[%s] %s%s", dsMo.getName(), isoPath, isoFileName), morDs);
-    }
-
     protected Answer execute(ReadyCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource ReadyCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource ReadyCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -3602,7 +2041,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(GetHostStatsCommand cmd) {
         if (s_logger.isTraceEnabled()) {
-            s_logger.trace("Executing resource GetHostStatsCommand: " + _gson.toJson(cmd));
+            s_logger.trace("Executing resource GetHostStatsCommand: " + gson.toJson(cmd));
         }
 
         VmwareContext context = getServiceContext();
@@ -3627,7 +2066,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         }
 
         if (s_logger.isTraceEnabled()) {
-            s_logger.trace("GetHostStats Answer: " + _gson.toJson(answer));
+            s_logger.trace("GetHostStats Answer: " + gson.toJson(answer));
         }
 
         return answer;
@@ -3635,7 +2074,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(GetVmStatsCommand cmd) {
         if (s_logger.isTraceEnabled()) {
-            s_logger.trace("Executing resource GetVmStatsCommand: " + _gson.toJson(cmd));
+            s_logger.trace("Executing resource GetVmStatsCommand: " + gson.toJson(cmd));
         }
 
         HashMap<String, VmStatsEntry> vmStatsMap = null;
@@ -3670,7 +2109,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         Answer answer = new GetVmStatsAnswer(cmd, vmStatsMap);
 
         if (s_logger.isTraceEnabled()) {
-            s_logger.trace("Report GetVmStatsAnswer: " + _gson.toJson(answer));
+            s_logger.trace("Report GetVmStatsAnswer: " + gson.toJson(answer));
         }
         return answer;
     }
@@ -3810,7 +2249,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
             for (String chainInfo : cmd.getVolumeUuids()) {
                 if (chainInfo != null) {
-                    VirtualMachineDiskInfo infoInChain = _gson.fromJson(chainInfo, VirtualMachineDiskInfo.class);
+                    VirtualMachineDiskInfo infoInChain = gson.fromJson(chainInfo, VirtualMachineDiskInfo.class);
                     if (infoInChain != null) {
                         String[] disks = infoInChain.getDiskChain();
                         if (disks.length > 0) {
@@ -3845,7 +2284,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(CheckHealthCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource CheckHealthCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource CheckHealthCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -3866,7 +2305,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(StopCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource StopCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource StopCommand: " + gson.toJson(cmd));
         }
 
         // In the stop command, we're passed in the name of the VM as seen by cloudstack,
@@ -3898,7 +2337,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
                         if (cmd.isForceStop()) {
                             success = vmMo.powerOff();
                         } else {
-                            success = vmMo.safePowerOff(_shutdownWaitMs);
+                            success = vmMo.safePowerOff(shutdownWaitMs);
                         }
                         if (!success) {
                             msg = "Have problem in powering off VM " + cmd.getVmName() + ", let the process continue";
@@ -3932,7 +2371,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(RebootRouterCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource RebootRouterCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource RebootRouterCommand: " + gson.toJson(cmd));
         }
 
         RebootAnswer answer = (RebootAnswer) execute((RebootCommand) cmd);
@@ -3951,7 +2390,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(RebootCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource RebootCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource RebootCommand: " + gson.toJson(cmd));
         }
 
         boolean toolsInstallerMounted = false;
@@ -4041,7 +2480,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(CheckVirtualMachineCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource CheckVirtualMachineCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource CheckVirtualMachineCommand: " + gson.toJson(cmd));
         }
 
         final String vmName = cmd.getVmName();
@@ -4074,7 +2513,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(PrepareForMigrationCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource PrepareForMigrationCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource PrepareForMigrationCommand: " + gson.toJson(cmd));
         }
 
         VirtualMachineTO vm = cmd.getVirtualMachine();
@@ -4107,12 +2546,12 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
                 prepareNetworkFromNicInfo(new HostMO(getServiceContext(), _morHyperHost), nic, false, cmd.getVirtualMachine().getType());
             }
 
-            List<Pair<String, Long>> secStoreUrlAndIdList = mgr.getSecondaryStorageStoresUrlAndIdList(Long.parseLong(_dcId));
+            List<Pair<String, Long>> secStoreUrlAndIdList = mgr.getSecondaryStorageStoresUrlAndIdList(Long.parseLong(dcId));
             for (Pair<String, Long> secStoreUrlAndId : secStoreUrlAndIdList) {
                 String secStoreUrl = secStoreUrlAndId.first();
                 Long secStoreId = secStoreUrlAndId.second();
                 if (secStoreUrl == null) {
-                    String msg = String.format("Secondary storage for dc %s is not ready yet?", _dcId);
+                    String msg = String.format("Secondary storage for dc %s is not ready yet?", dcId);
                     throw new Exception(msg);
                 }
 
@@ -4143,7 +2582,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         if (s_logger.isInfoEnabled()) {
             s_logger.info(String.format("excuting MigrateVmToPoolCommand %s -> %s", cmd.getVmName(), cmd.getDestinationPool()));
             if (s_logger.isDebugEnabled()) {
-                s_logger.debug("MigrateVmToPoolCommand: " + _gson.toJson(cmd));
+                s_logger.debug("MigrateVmToPoolCommand: " + gson.toJson(cmd));
             }
         }
 
@@ -4224,7 +2663,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
                 VirtualMachineDiskInfo diskInfo = diskInfoBuilder.getDiskInfoByBackingFileBaseName(newPath, poolUuid);
                 newVol.setId(volumeDeviceKey.get(disk.getKey()));
                 newVol.setPath(newPath);
-                newVol.setChainInfo(_gson.toJson(diskInfo));
+                newVol.setChainInfo(gson.toJson(diskInfo));
                 volumeToList.add(newVol);
             }
             return new MigrateVmToPoolAnswer((MigrateVmToPoolCommand) cmd, volumeToList);
@@ -4301,7 +2740,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(MigrateCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource MigrateCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource MigrateCommand: " + gson.toJson(cmd));
         }
 
         final String vmName = cmd.getVmName();
@@ -4344,7 +2783,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
     protected Answer execute(MigrateWithStorageCommand cmd) {
 
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource MigrateWithStorageCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource MigrateWithStorageCommand: " + gson.toJson(cmd));
         }
 
         VirtualMachineTO vmTo = cmd.getVirtualMachine();
@@ -4492,12 +2931,12 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
             }
 
             // Ensure all secondary storage mounted on target host
-            List<Pair<String, Long>> secStoreUrlAndIdList = mgr.getSecondaryStorageStoresUrlAndIdList(Long.parseLong(_dcId));
+            List<Pair<String, Long>> secStoreUrlAndIdList = mgr.getSecondaryStorageStoresUrlAndIdList(Long.parseLong(dcId));
             for (Pair<String, Long> secStoreUrlAndId : secStoreUrlAndIdList) {
                 String secStoreUrl = secStoreUrlAndId.first();
                 Long secStoreId = secStoreUrlAndId.second();
                 if (secStoreUrl == null) {
-                    String msg = String.format("Secondary storage for dc %s is not ready yet?", _dcId);
+                    String msg = String.format("Secondary storage for dc %s is not ready yet?", dcId);
                     throw new Exception(msg);
                 }
 
@@ -4561,7 +3000,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
                         VirtualMachineDiskInfo diskInfo = diskInfoBuilder.getDiskInfoByBackingFileBaseName(newPath, poolName);
                         newVol.setId(volumeId);
                         newVol.setPath(newPath);
-                        newVol.setChainInfo(_gson.toJson(diskInfo));
+                        newVol.setChainInfo(gson.toJson(diskInfo));
                         volumeToList.add(newVol);
                         break;
                     }
@@ -4733,7 +3172,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         StorageFilerTO poolTo = cmd.getPool();
 
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource MigrateVolumeCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource MigrateVolumeCommand: " + gson.toJson(cmd));
         }
 
         String vmName = cmd.getAttachedVmName();
@@ -4831,7 +3270,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
                     }
             }
             VirtualMachineDiskInfoBuilder diskInfoBuilder = vmMo.getDiskInfoBuilder();
-            String chainInfo = _gson.toJson(diskInfoBuilder.getDiskInfoByBackingFileBaseName(volumePath, poolTo.getUuid().replace("-", "")));
+            String chainInfo = gson.toJson(diskInfoBuilder.getDiskInfoByBackingFileBaseName(volumePath, poolTo.getUuid().replace("-", "")));
             MigrateVolumeAnswer answer = new MigrateVolumeAnswer(cmd, true, null, volumePath);
             answer.setVolumeChainInfo(chainInfo);
             return answer;
@@ -4842,7 +3281,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         }
     }
 
-    private Pair<VirtualDisk, String> getVirtualDiskInfo(VirtualMachineMO vmMo, String srcDiskName) throws Exception {
+    Pair<VirtualDisk, String> getVirtualDiskInfo(VirtualMachineMO vmMo, String srcDiskName) throws Exception {
         Pair<VirtualDisk, String> deviceInfo = vmMo.getDiskDevice(srcDiskName);
         if (deviceInfo == null) {
             throw new Exception("No such disk device: " + srcDiskName);
@@ -4918,7 +3357,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(ModifyStoragePoolCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource ModifyStoragePoolCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource ModifyStoragePoolCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -4978,7 +3417,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(DeleteStoragePoolCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource DeleteStoragePoolCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource DeleteStoragePoolCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -5021,7 +3460,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected AttachIsoAnswer execute(AttachIsoCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource AttachIsoCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource AttachIsoCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -5140,7 +3579,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(ValidateSnapshotCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource ValidateSnapshotCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource ValidateSnapshotCommand: " + gson.toJson(cmd));
         }
 
         // the command is no longer available
@@ -5153,7 +3592,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(ManageSnapshotCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource ManageSnapshotCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource ManageSnapshotCommand: " + gson.toJson(cmd));
         }
 
         long snapshotId = cmd.getSnapshotId();
@@ -5185,7 +3624,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(BackupSnapshotCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource BackupSnapshotCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource BackupSnapshotCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -5242,7 +3681,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(CreateVolumeFromSnapshotCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource CreateVolumeFromSnapshotCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource CreateVolumeFromSnapshotCommand: " + gson.toJson(cmd));
         }
 
         String details = null;
@@ -5268,7 +3707,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(CreatePrivateTemplateFromVolumeCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource CreatePrivateTemplateFromVolumeCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource CreatePrivateTemplateFromVolumeCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -5295,7 +3734,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(CreatePrivateTemplateFromSnapshotCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource CreatePrivateTemplateFromSnapshotCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource CreatePrivateTemplateFromSnapshotCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -5316,7 +3755,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(GetStorageStatsCommand cmd) {
         if (s_logger.isTraceEnabled()) {
-            s_logger.trace("Executing resource GetStorageStatsCommand: " + _gson.toJson(cmd));
+            s_logger.trace("Executing resource GetStorageStatsCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -5365,7 +3804,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(GetVncPortCommand cmd) {
         if (s_logger.isTraceEnabled()) {
-            s_logger.trace("Executing resource GetVncPortCommand: " + _gson.toJson(cmd));
+            s_logger.trace("Executing resource GetVncPortCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -5407,7 +3846,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(SetupCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource SetupCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource SetupCommand: " + gson.toJson(cmd));
         }
 
         return new SetupAnswer(cmd, false);
@@ -5415,7 +3854,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(MaintainCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource MaintainCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource MaintainCommand: " + gson.toJson(cmd));
         }
 
         return new MaintainAnswer(cmd, "Put host in maintaince");
@@ -5423,7 +3862,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(PingTestCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource PingTestCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource PingTestCommand: " + gson.toJson(cmd));
         }
 
         String controlIp = cmd.getRouterIp();
@@ -5467,7 +3906,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(CheckOnHostCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource CheckOnHostCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource CheckOnHostCommand: " + gson.toJson(cmd));
         }
 
         return new CheckOnHostAnswer(cmd, null, "Not Implmeneted");
@@ -5484,7 +3923,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(GetVmIpAddressCommand cmd) {
         if (s_logger.isTraceEnabled()) {
-            s_logger.trace("Executing resource command GetVmIpAddressCommand: " + _gson.toJson(cmd));
+            s_logger.trace("Executing resource command GetVmIpAddressCommand: " + gson.toJson(cmd));
         }
 
         String details = "Unable to find IP Address of VM. ";
@@ -5530,7 +3969,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
         answer = new Answer(cmd, result, details);
         if (s_logger.isTraceEnabled()) {
-            s_logger.trace("Returning GetVmIpAddressAnswer: " + _gson.toJson(answer));
+            s_logger.trace("Returning GetVmIpAddressAnswer: " + gson.toJson(answer));
         }
         return answer;
     }
@@ -5538,7 +3977,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
     @Override
     public PrimaryStorageDownloadAnswer execute(PrimaryStorageDownloadCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource PrimaryStorageDownloadCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource PrimaryStorageDownloadCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -5567,7 +4006,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     protected Answer execute(UnregisterVMCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource UnregisterVMCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource UnregisterVMCommand: " + gson.toJson(cmd));
         }
 
         VmwareContext context = getServiceContext();
@@ -5617,7 +4056,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
      * @return
      */
     protected Answer execute(UnregisterNicCommand cmd) {
-        s_logger.info("Executing resource UnregisterNicCommand: " + _gson.toJson(cmd));
+        s_logger.info("Executing resource UnregisterNicCommand: " + gson.toJson(cmd));
 
         if (_guestTrafficInfo == null) {
             return new Answer(cmd, false, "No Guest Traffic Info found, unable to determine where to clean up");
@@ -5676,7 +4115,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
     @Override
     public CopyVolumeAnswer execute(CopyVolumeCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource CopyVolumeCommand: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource CopyVolumeCommand: " + gson.toJson(cmd));
         }
 
         try {
@@ -5869,11 +4308,11 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
                     cmd.setPoolInfo(pInfo);
                     cmd.setGuid(poolUuid); // give storage host the same UUID as the local storage pool itself
                     cmd.setResourceType(Storage.StorageResourceType.STORAGE_POOL);
-                    cmd.setDataCenter(_dcId);
+                    cmd.setDataCenter(dcId);
                     cmd.setPod(_pod);
                     cmd.setCluster(_cluster);
 
-                    s_logger.info("Add local storage startup command: " + _gson.toJson(cmd));
+                    s_logger.info("Add local storage startup command: " + gson.toJson(cmd));
                     storageCmds.add(cmd);
                 }
 
@@ -5917,7 +4356,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         cmd.setHostDetails(details);
         cmd.setName(_url);
         cmd.setGuid(_guid);
-        cmd.setDataCenter(_dcId);
+        cmd.setDataCenter(dcId);
         cmd.setIqn(getIqn());
         cmd.setPod(_pod);
         cmd.setCluster(_cluster);
@@ -5955,7 +4394,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
         VmwareHypervisorHostResourceSummary summary = hyperHost.getHyperHostResourceSummary();
 
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Startup report on host hardware info. " + _gson.toJson(summary));
+            s_logger.info("Startup report on host hardware info. " + gson.toJson(summary));
         }
 
         cmd.setCaps("hvm");
@@ -5980,7 +4419,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
             }
 
             if (s_logger.isInfoEnabled()) {
-                s_logger.info("Startup report on host network info. " + _gson.toJson(summary));
+                s_logger.info("Startup report on host network info. " + gson.toJson(summary));
             }
 
             cmd.setPrivateIpAddress(summary.getHostIp());
@@ -6085,35 +4524,6 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
                 s_logger.error("Unexpected exception ", e);
             }
         }
-    }
-
-    private VirtualMachineGuestOsIdentifier translateGuestOsIdentifier(String cpuArchitecture, String guestOs, String cloudGuestOs) {
-        if (cpuArchitecture == null) {
-            s_logger.warn("CPU arch is not set, default to i386. guest os: " + guestOs);
-            cpuArchitecture = "i386";
-        }
-
-        if (cloudGuestOs == null) {
-            s_logger.warn("Guest OS mapping name is not set for guest os: " + guestOs);
-        }
-
-        VirtualMachineGuestOsIdentifier identifier = null;
-        try {
-            if (cloudGuestOs != null) {
-                identifier = VirtualMachineGuestOsIdentifier.fromValue(cloudGuestOs);
-                s_logger.debug("Using mapping name : " + identifier.toString());
-            }
-        } catch (IllegalArgumentException e) {
-            s_logger.warn("Unable to find Guest OS Identifier in VMware for mapping name: " + cloudGuestOs + ". Continuing with defaults.");
-        }
-        if (identifier != null) {
-            return identifier;
-        }
-
-        if (cpuArchitecture.equalsIgnoreCase("x86_64")) {
-            return VirtualMachineGuestOsIdentifier.OTHER_GUEST_64;
-        }
-        return VirtualMachineGuestOsIdentifier.OTHER_GUEST;
     }
 
     private HashMap<String, HostVmStateReportEntry> getHostVmStateReport() throws Exception {
@@ -6553,7 +4963,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
             _url = (String) params.get("url");
             _username = (String) params.get("username");
             _password = (String) params.get("password");
-            _dcId = (String) params.get("zone");
+            dcId = (String) params.get("zone");
             _pod = (String) params.get("pod");
             _cluster = (String) params.get("cluster");
 
@@ -6592,7 +5002,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
             if (_guestTrafficInfo.getVirtualSwitchType() == VirtualSwitchType.NexusDistributedVirtualSwitch
                     || _publicTrafficInfo.getVirtualSwitchType() == VirtualSwitchType.NexusDistributedVirtualSwitch) {
-                _privateNetworkVSwitchName = mgr.getPrivateVSwitchName(Long.parseLong(_dcId), HypervisorType.VMware);
+                _privateNetworkVSwitchName = mgr.getPrivateVSwitchName(Long.parseLong(dcId), HypervisorType.VMware);
                 _vsmCredentials = mgr.getNexusVSMCredentialsByClusterId(Long.parseLong(_cluster));
             }
 
@@ -6629,16 +5039,16 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
             boolObj = (Boolean) params.get("vm.instancename.flag");
             if (boolObj != null && boolObj.booleanValue()) {
-                _instanceNameFlag = true;
+                instanceNameFlag = true;
             } else {
-                _instanceNameFlag = false;
+                instanceNameFlag = false;
             }
 
             value = (String) params.get("scripts.timeout");
             int timeout = NumbersUtil.parseInt(value, 1440) * 1000;
 
             storageNfsVersion = NfsSecondaryStorageResource.retrieveNfsVersionFromParams(params);
-            _storageProcessor = new VmwareStorageProcessor((VmwareHostService) this, _fullCloneFlag, (VmwareStorageMount) mgr, timeout, this, _shutdownWaitMs, null,
+            _storageProcessor = new VmwareStorageProcessor((VmwareHostService) this, _fullCloneFlag, (VmwareStorageMount) mgr, timeout, this, shutdownWaitMs, null,
                     storageNfsVersion);
             storageHandler = new VmwareStorageSubsystemCommandHandler(_storageProcessor, storageNfsVersion);
 
@@ -6795,7 +5205,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
     @Override
     public Answer execute(DestroyCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource DestroyCommand to evict template from storage pool: " + _gson.toJson(cmd));
+            s_logger.info("Executing resource DestroyCommand to evict template from storage pool: " + gson.toJson(cmd));
         }
 
         try {
@@ -7150,7 +5560,7 @@ public class VmwareResource implements StoragePoolResource, ServerResource, Vmwa
 
     private Answer execute(GetUnmanagedInstancesCommand cmd) {
         if (s_logger.isInfoEnabled()) {
-            s_logger.info("Executing resource GetUnmanagedInstancesCommand " + _gson.toJson(cmd));
+            s_logger.info("Executing resource GetUnmanagedInstancesCommand " + gson.toJson(cmd));
         }
 
         VmwareContext context = getServiceContext();
