+++ plugins/hypervisors/vmware/src/test/java/com/cloud/hypervisor/vmware/resource/VmwareResourceTest.java
@@ -55,7 +55,6 @@ import com.cloud.agent.api.ScaleVmAnswer;
 import com.cloud.agent.api.ScaleVmCommand;
 import com.cloud.agent.api.to.DataTO;
 import com.cloud.agent.api.to.NfsTO;
-import com.cloud.agent.api.to.NicTO;
 import com.cloud.agent.api.to.VirtualMachineTO;
 import com.cloud.agent.api.to.VolumeTO;
 import com.cloud.hypervisor.Hypervisor.HypervisorType;
@@ -212,24 +211,6 @@ public class VmwareResourceTest {
 
         _resource.execute(cmd);
         verify(_resource).execute(cmd);
-    }
-
-    @Test
-    public void testGenerateMacSequence() {
-        final NicTO nicTo1 = new NicTO();
-        nicTo1.setMac("01:23:45:67:89:AB");
-        nicTo1.setDeviceId(1);
-
-        final NicTO nicTo2 = new NicTO();
-        nicTo2.setMac("02:00:65:b5:00:03");
-        nicTo2.setDeviceId(0);
-
-        //final NicTO [] nicTOs = {nicTO1, nicTO2, nicTO3};
-        //final NicTO[] nics = new NicTO[]{nic};
-        final NicTO[] nics = new NicTO[] {nicTo1, nicTo2};
-
-        String macSequence = _resource.generateMacSequence(nics);
-        assertEquals(macSequence, "02:00:65:b5:00:03|01:23:45:67:89:AB");
     }
 
     @Test
